rename: AsciiExportModal → TextExportModal (#208)
The modal handles three formats (ASCII, Unicode half-blocks, Unicode
full-blocks) — the "Ascii" name leaked one specific case. "TextExport"
matches the modal title ("Export as text") and the conceptual scope.

Renames the file, the import in QRCodeCreate.vue, and the
reactive state (isTextExportModalOpen, openTextExportModal). Also fixes
two stale assertions in convertToText.test.ts that were still expecting
the pre-inversion ASCII glyph order.
