+++ locales/af.json
@@ -51,11 +51,11 @@
   "Last saved locally": "Gelaai vanaf plaaslike berging",
   "{count} piece(s) of data detected": "{count} Gedetekteerde datastuk(ke)",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-kodes is geskep.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bis-karakters — werk in gewone teks, e-pos en terminals.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bis, omgekeerde polariteit (# is agtergrond) — werk in gewone teks, e-pos en terminals.",
   "Add frame": "Voeg raam by",
   "Address": "Adres",
-  "ASCII": "ASCII",
+  "ASCII": "ASKII",
   "Available Fields": "Beskikbare velde",
   "Background color": "Agtergrondkleur",
   "Batch export": "Lotuitvoer",
@@ -82,10 +82,10 @@
   "Close dialog": "Sluit dialoog",
   "Close guide": "Sluit gids",
   "Close scanner": "Sluit skandeerder",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompak: twee rye in een lyn saamgepers met ▀ ▄ █-karakters.",
   "Configure frame settings": "Konfigureer raaminstellings",
   "Contact Card": "Kontakkaart",
-  "Copied": "Copied",
+  "Copied": "Gekopieer",
   "Copy QR Code to clipboard": "Kopieer QR-kode na die knipbord",
   "Copy to clipboard": "Kopieer na knipbord",
   "Corners Dot color": "Hoeke Dot kleur",
@@ -113,10 +113,10 @@
   "dots": "stippels",
   "Dots color": "Punte kleur",
   "Dots type": "Punte-tipe",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Laai af .md",
+  "Download .txt": "Laai .txt af",
+  "Download all as .md": "Laai alles af as .md",
+  "Download all as .txt": "Laai alles af as .txt",
   "Download QR Code as JPG": "Laai QR-kode af as JPG",
   "Download QR Code as PNG": "Laai QR-kode af as PNG",
   "Download QR Code as SVG": "Laai QR-kode af as SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Voorbeeldlêer (vCard)",
   "Example with Custom Filenames": "Voorbeeld met pasgemaakte lêernaam",
   "Export": "Uitvoer",
-  "Export as text": "Export as text",
+  "Export as text": "Eksporteer as teks",
   "Export QR code": "Eksporteer QR-kode",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksporteer QR-kode as ASCII- of Unicode-teks",
   "extra-rounded": "ekstra-afgerond",
   "Failed to copy to clipboard": "Het misluk om na die knipbord te kopieer",
   "Failed to load changelog": "Misluk om die veranderinglog te laai",
@@ -161,12 +161,12 @@
   "First Name": "Voorname",
   "First row preview:": "Eerste ry voorskou:",
   "Font family": "Letterfamilie",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Raametiket word uitgesluit van die teks-uitvoer — die ASCII-uitset bevat slegs die QR-modules.",
   "Frame preset": "Kader-voorsetting",
   "Frame settings": "Raaminstellings",
   "Frame style": "Raamstyl",
   "Frame text": "Raamteks",
-  "Generating": "Generating",
+  "Generating": "Genererend",
   "GitHub repository for this project": "GitHub-bewaarplek vir hierdie projek",
   "Have a beautiful day!": "Het 'n pragtige dag!",
   "Have nice day!": "Geniet jou dag!",
@@ -204,7 +204,7 @@
   "Message": "Boodskap",
   "Mini QR Code Generator": "Mini QR-kode-generator",
   "MiniQR's QR rendering engine has been updated": "MiniQR se QR-weergawemotor is opgedateer.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Die meeste QR-skandeerders sal hierdie formaat nie lees nie omdat die #-glyf gapings tussen rye laat. Gebruik Unicode (half-/vol blokke) vir 'n skandeerbare uitvoer.",
   "Must be between 0 and 1": "Moet tussen 0 en 1 wees",
   "Network Name (SSID)": "Netwerknaam (SSID)",
   "Network password (case-sensitive)": "Netwerkwagwoord (lettergrootheidsgevoelig)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Laai asseblief eers 'n lêer op.",
   "Position": "Posisie",
   "Preset": "Voorsetting",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Voorskou van die eerste ry",
   "Processing...": "Verwerking...",
   "QR Code Content": "QR-kode-inhoud",
   "QR code preset": "QR-kode-voorsinstelling",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Eenvoudige URL/Teks",
   "Single export": "Enkel uitvoer",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Vaste blokkarakterte (██) — beste getrouheid in moderne lettertipes.",
   "Sponsor": "Borg",
   "square": "vierkant",
   "Start Date": "Begindatum",
@@ -291,15 +291,15 @@
   "Text": "Teks",
   "Text color": "Tekskleur",
   "Text is required": "Teks is vereis",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Tekstuitset sluit die raam en die beeld uit.",
   "Text position": "Teksposisie",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Daar is te veel QR-kodes om gelyktydig na die knipbord te kopieer. Laai dit asseblief eerder as SVG of PNG af.",
   "Tip: For best results, use a clear image with good lighting.": "Tip: Vir die beste resultate, gebruik 'n duidelike beeld met goeie beligting.",
   "Title": "Titel",
   "Toggle dark mode": "Skakel donker modus aan of af",
   "top": "top",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (volledige blokke)",
+  "Unicode (half-blocks)": "Unicode (halfblokke)",
   "Upload a CSV file": "Laai 'n CSV-lêer op",
   "Upload image": "Laai beeld op",
   "Upload QR Code Image": "Laai QR-kode-prent op",
