+++ locales/bg.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Зареждане от локално хранилище",
   "{count} piece(s) of data detected": "{count} част(и) от откритите данни",
   "{index} / {count} QR codes have been created.": "{index} / {count} Създадени са QR кодове.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-битови символи — работят в обикновен текст, електронна поща и терминали.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-битов, с обърната полярност (# е фон) — работи в обикновен текст, имейли и терминали.",
   "Add frame": "Добавяне на рамка",
   "Address": "Адрес",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Затваряне на диалоговия прозорец",
   "Close guide": "Затвори ръководство",
   "Close scanner": "Затваряне на скенера",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Компактно: два реда, събрани в един ред с помощта на символите ▀ ▄ █.",
   "Configure frame settings": "Конфигуриране на настройките на рамката",
   "Contact Card": "Контактна карта",
-  "Copied": "Copied",
+  "Copied": "Копирано",
   "Copy QR Code to clipboard": "Копиране на QR кода в клипборда",
   "Copy to clipboard": "Копирай в клипборда",
   "Corners Dot color": "Ъгли Цвят на точката",
@@ -113,10 +113,10 @@
   "dots": "точки",
   "Dots color": "Цвят на точките",
   "Dots type": "Тип точки",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Изтегли .md",
+  "Download .txt": "Изтегли .txt",
+  "Download all as .md": "Изтегли всичко като .md",
+  "Download all as .txt": "Изтегли всичко като .txt",
   "Download QR Code as JPG": "Изтегляне на QR код като JPG",
   "Download QR Code as PNG": "Изтегляне на QR код като PNG",
   "Download QR Code as SVG": "Изтегляне на QR код като SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Примерен файл (vCard)",
   "Example with Custom Filenames": "Пример с персонализирани имена на файлове",
   "Export": "Износ",
-  "Export as text": "Export as text",
+  "Export as text": "Експортирай като текст",
   "Export QR code": "Експортиране на QR код",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Експортиране на QR кода като ASCII или Unicode текст",
   "extra-rounded": "изключително закръглени",
   "Failed to copy to clipboard": "Неуспешно копиране в клипборда",
   "Failed to load changelog": "Неуспешно зареждане на списъка с промените",
@@ -161,12 +161,12 @@
   "First Name": "Първо име",
   "First row preview:": "Предварителен преглед на първия ред:",
   "Font family": "Семейство шрифтове",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Етикетът на рамката не се включва в експортирания текст — ASCII файлът съдържа само QR модулите.",
   "Frame preset": "Предварителна настройка на кадъра",
   "Frame settings": "Настройки на рамката",
   "Frame style": "Стил на рамката",
   "Frame text": "Текст на рамката",
-  "Generating": "Generating",
+  "Generating": "Генериране",
   "GitHub repository for this project": "Хранилище на GitHub за този проект",
   "Have a beautiful day!": "Пожелавам ви прекрасен ден!",
   "Have nice day!": "Хубав ден!",
@@ -204,7 +204,7 @@
   "Message": "Съобщение",
   "Mini QR Code Generator": "Генератор на мини QR код",
   "MiniQR's QR rendering engine has been updated": "Двигателят за визуализация на QR кодове на MiniQR беше актуализиран",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Повечето QR-скенери не могат да прочетат този формат, тъй като символите # оставят празнини между редовете. За да получите файл, подходящ за сканиране, използвайте Unicode (половинни/пълни блокове).",
   "Must be between 0 and 1": "Трябва да е число между 0 и 1",
   "Network Name (SSID)": "Име на мрежата (SSID)",
   "Network password (case-sensitive)": "Мрежова парола (чувствителна към големи и малки букви)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Моля, първо качете файл",
   "Position": "Позиция",
   "Preset": "Предварителна настройка",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Преглед на първия ред",
   "Processing...": "Обработка...",
   "QR Code Content": "Съдържание на QR кода",
   "QR code preset": "Предварително зададен QR код",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Обикновен URL/текст",
   "Single export": "Единичен износ",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Плътни блокови символи (██) — най-добрата прецизност при съвременните шрифтове.",
   "Sponsor": "Спонсор",
   "square": "квадрат",
   "Start Date": "Начална дата",
@@ -291,15 +291,15 @@
   "Text": "Текст",
   "Text color": "Цвят на текста",
   "Text is required": "Изисква се текст",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Текстът не включва рамката и изображението",
   "Text position": "Позиция на текста",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Има твърде много QR кодове, които могат да бъдат копирани едновременно в клипборда. Вместо това ги изтеглете като SVG или PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Съвет: За най-добри резултати използвайте ясно изображение с добро осветление.",
   "Title": "Заглавие",
   "Toggle dark mode": "Превключване на тъмен режим",
   "top": "топ",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Юникод (пълни блокове)",
+  "Unicode (half-blocks)": "Юникод (полублокове)",
   "Upload a CSV file": "Качване на CSV файл",
   "Upload image": "Качване на изображение",
   "Upload QR Code Image": "Качване на изображение на QR код",
