+++ locales/bn.json
@@ -51,11 +51,11 @@
   "Last saved locally": "সর্বশেষ লোকালি সেভ করা",
   "{count} piece(s) of data detected": "{count} টি ডাটা শনাক্ত হয়েছে",
   "{index} / {count} QR codes have been created.": "{index} / {count} টি কিউআর কোড তৈরি হয়েছে।",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "৭-বিট অক্ষর — প্লেন টেক্সট, ইমেইল এবং টার্মিনালে কাজ করে।",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "৭-বিট, বিপরীত পোলারিটি (# হল পটভূমি) — প্লেন টেক্সট, ইমেইল এবং টার্মিনালে কাজ করে।",
   "Add frame": "ফ্রেম যোগ করুন",
   "Address": "ঠিকানা",
-  "ASCII": "ASCII",
+  "ASCII": "অ্যাস্কি",
   "Available Fields": "উপলব্ধ ফিল্ড",
   "Background color": "ব্যাকগ্রাউন্ডের রং",
   "Batch export": "ব্যাচ এক্সপোর্ট",
@@ -82,10 +82,10 @@
   "Close dialog": "ডায়ালগ বন্ধ করুন",
   "Close guide": "গাইড বন্ধ করুন",
   "Close scanner": "স্ক্যানার বন্ধ করুন",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "কমপ্যাক্ট: ▀ ▄ █ অক্ষর ব্যবহার করে দুই সারিকে এক সারিতে সঙ্কুচিত করা।",
   "Configure frame settings": "ফ্রেম সেটিংস কনফিগার করুন",
   "Contact Card": "কনট্যাক্ট কার্ড",
-  "Copied": "Copied",
+  "Copied": "নকল করা হয়েছে",
   "Copy QR Code to clipboard": "কিউআর কোড ক্লিপবোর্ডে কপি করুন",
   "Copy to clipboard": "ক্লিপবোর্ডে কপি করুন",
   "Corners Dot color": "কোণের ডটের রং",
@@ -113,10 +113,10 @@
   "dots": "ডটস",
   "Dots color": "ডটের রং",
   "Dots type": "ডটের ধরন",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "ডাউনলোড .md",
+  "Download .txt": "ডাউনলোড .txt",
+  "Download all as .md": "সবকিছু .md ফাইল হিসেবে ডাউনলোড করুন",
+  "Download all as .txt": "সবকিছু .txt হিসেবে ডাউনলোড করুন",
   "Download QR Code as JPG": "কিউআর কোড JPG হিসেবে ডাউনলোড করুন",
   "Download QR Code as PNG": "কিউআর কোড PNG হিসেবে ডাউনলোড করুন",
   "Download QR Code as SVG": "কিউআর কোড SVG হিসেবে ডাউনলোড করুন",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "উদাহরণ ফাইল (vCard)",
   "Example with Custom Filenames": "কাস্টম ফাইলনেমসহ উদাহরণ",
   "Export": "এক্সপোর্ট",
-  "Export as text": "Export as text",
+  "Export as text": "টেক্সট হিসেবে রপ্তানি করুন",
   "Export QR code": "কিউআর কোড এক্সপোর্ট করুন",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR কোড ASCII বা Unicode টেক্সট হিসেবে রপ্তানি করুন",
   "extra-rounded": "এক্সট্রা-রাউন্ডেড",
   "Failed to copy to clipboard": "ক্লিপবোর্ডে কপি করতে ব্যর্থ",
   "Failed to load changelog": "চেঞ্জলগ লোড করতে ব্যর্থ",
@@ -161,12 +161,12 @@
   "First Name": "প্রথম নাম",
   "First row preview:": "প্রথম সারির প্রিভিউ:",
   "Font family": "ফন্ট পরিবার",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "ফ্রেম লেবেল টেক্সট রপ্তানি থেকে বাদ দেওয়া হয়েছে — ASCII আউটপুটে শুধুমাত্র QR মডিউলগুলো রয়েছে।",
   "Frame preset": "ফ্রেম প্রিসেট",
   "Frame settings": "ফ্রেম সেটিংস",
   "Frame style": "ফ্রেম স্টাইল",
   "Frame text": "ফ্রেম টেক্সট",
-  "Generating": "Generating",
+  "Generating": "উৎপাদন",
   "GitHub repository for this project": "এই প্রকল্পের GitHub রেপোজিটরি",
   "Have a beautiful day!": "একটি সুন্দর দিন কাটুক!",
   "Have nice day!": "সুন্দর দিন কাটুক!",
@@ -204,7 +204,7 @@
   "Message": "বার্তা",
   "Mini QR Code Generator": "মিনি কিউআর কোড জেনারেটর",
   "MiniQR's QR rendering engine has been updated": "MiniQR-এর QR রেন্ডারিং ইঞ্জিন আপডেট করা হয়েছে।",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "অধিকাংশ QR স্ক্যানার এই ফরম্যাট পড়তে পারবে না কারণ # গ্লিফগুলো সারিগুলোর মধ্যে ফাঁক রেখে যায়। স্ক্যানযোগ্য এক্সপোর্টের জন্য ইউনিকোড (আধা/পূর্ণ ব্লক) ব্যবহার করুন।",
   "Must be between 0 and 1": "0 থেকে 1 এর মধ্যে হতে হবে",
   "Network Name (SSID)": "নেটওয়ার্কের নাম (SSID)",
   "Network password (case-sensitive)": "নেটওয়ার্ক পাসওয়ার্ড (কেস-সেনসিটিভ)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "প্রথমে একটি ফাইল আপলোড করুন",
   "Position": "অবস্থান",
   "Preset": "প্রিসেট",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "প্রথম সারির পূর্বরূপ",
   "Processing...": "প্রসেসিং...",
   "QR Code Content": "কিউআর কোডের কনটেন্ট",
   "QR code preset": "কিউআর কোড প্রিসেট",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "সাধারণ URL/টেক্সট",
   "Single export": "সিঙ্গেল এক্সপোর্ট",
   "SMS": "এসএমএস",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "সলিড ব্লক অক্ষর (██) — আধুনিক ফন্টে সেরা বিশ্বস্ততা।",
   "Sponsor": "স্পনসর",
   "square": "স্কোয়ার",
   "Start Date": "শুরুর তারিখ",
@@ -291,15 +291,15 @@
   "Text": "টেক্সট",
   "Text color": "টেক্সটের রং",
   "Text is required": "টেক্সট প্রয়োজন",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "টেক্সট আউটপুটে ফ্রেম এবং চিত্র অন্তর্ভুক্ত হয় না।",
   "Text position": "টেক্সটের অবস্থান",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "একসাথে এতগুলো কিউআর কোড ক্লিপবোর্ডে কপি করা সম্ভব নয়। অনুগ্রহ করে SVG বা PNG হিসেবে ডাউনলোড করুন।",
   "Tip: For best results, use a clear image with good lighting.": "টিপস: ভালো ফলাফলের জন্য পরিষ্কার ও ভালো আলোযুক্ত ইমেজ ব্যবহার করুন।",
   "Title": "পদবী",
   "Toggle dark mode": "ডার্ক মোড চালু/বন্ধ",
   "top": "উপরে",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "ইউনিকোড (পূর্ণ ব্লক)",
+  "Unicode (half-blocks)": "ইউনিকোড (আধা-ব্লক)",
   "Upload a CSV file": "একটি CSV ফাইল আপলোড করুন",
   "Upload image": "ইমেজ আপলোড করুন",
   "Upload QR Code Image": "কিউআর কোড ইমেজ আপলোড করুন",
