+++ locales/ca.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Cargat des de l'emmagatzematge local",
   "{count} piece(s) of data detected": "{count} Peces de dades detectades",
   "{index} / {count} QR codes have been created.": "{index} / {count} S'han creat codis QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Caràcters de 7 bits — funciona en text pla, correu electrònic i terminals.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bits, polaritat invertida (# és el fons) — funciona en text pla, correu electrònic i terminals.",
   "Add frame": "Afegeix marc",
   "Address": "Adreça",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Tanca el diàleg",
   "Close guide": "Tanca la guia",
   "Close scanner": "Tanca l'escàner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compacte: dues files comprimides en una sola línia mitjançant els caràcters ▀ ▄ █.",
   "Configure frame settings": "Configurar la configuració del marc",
   "Contact Card": "Visita de contacte",
-  "Copied": "Copied",
+  "Copied": "Copiat",
   "Copy QR Code to clipboard": "Copia el codi QR al porta-retalls",
   "Copy to clipboard": "Còpia al porta-retalls",
   "Corners Dot color": "Punts de cantonada Dot color",
@@ -113,10 +113,10 @@
   "dots": "punts",
   "Dots color": "Color de punts",
   "Dots type": "Tipus de punts",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Descarrega .md",
+  "Download .txt": "Descarrega el .txt",
+  "Download all as .md": "Descarrega-ho tot com a .md",
+  "Download all as .txt": "Descarrega-ho tot com a .txt",
   "Download QR Code as JPG": "Descarrega el codi QR com a JPG",
   "Download QR Code as PNG": "Descarrega el codi QR com a PNG",
   "Download QR Code as SVG": "Descarrega el codi QR com a SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Fitxer d'exemple (vCard)",
   "Example with Custom Filenames": "Exemple amb noms de fitxer personalitzats",
   "Export": "Exportar",
-  "Export as text": "Export as text",
+  "Export as text": "Exporta com a text",
   "Export QR code": "Exporta el codi QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportar codi QR com a text ASCII o Unicode",
   "extra-rounded": "extra-arrodonit",
   "Failed to copy to clipboard": "No s'ha pogut copiar al porta-retalls",
   "Failed to load changelog": "S'ha fallat la càrrega del registre de canvis",
@@ -161,12 +161,12 @@
   "First Name": "Nom",
   "First row preview:": "Previsualització de la primera fila:",
   "Font family": "Família de tipus",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "L'etiqueta del marc està exclosa de l'exportació de text — la sortida ASCII només conté els mòduls QR.",
   "Frame preset": "Preajust de marc",
   "Frame settings": "Configuració del marc",
   "Frame style": "Estil de marc",
   "Frame text": "Marc de text",
-  "Generating": "Generating",
+  "Generating": "Generant",
   "GitHub repository for this project": "Repositori de GitHub per a aquest projecte",
   "Have a beautiful day!": "Que tinguis un bon dia!",
   "Have nice day!": "Que tinguis un bon dia!",
@@ -204,7 +204,7 @@
   "Message": "Missatge",
   "Mini QR Code Generator": "Generador de codis QR miniatura",
   "MiniQR's QR rendering engine has been updated": "El motor de renderització de QR de MiniQR s'ha actualitzat.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "La majoria de lectors de QR no llegiran aquest format perquè els glifs # deixen espais entre les línies. Utilitzeu Unicode (blocs mitjans/complets) per a una exportació escanejable.",
   "Must be between 0 and 1": "Ha de ser entre 0 i 1",
   "Network Name (SSID)": "Nom de la xarxa (SSID)",
   "Network password (case-sensitive)": "Contrasenya de la xarxa (amb distinció de majúscules i minúscules)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Si us plau, pugeu primer un fitxer.",
   "Position": "Posició",
   "Preset": "Preajust",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Previsualització de la primera fila",
   "Processing...": "Processament...",
   "QR Code Content": "Contingut del codi QR",
   "QR code preset": "Preajust de codi QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/Text senzill",
   "Single export": "Exportació única",
   "SMS": "Missatge de text",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Caràcters de bloc sòlid (██) — la millor fidelitat en les tipografies modernes.",
   "Sponsor": "Patrocinador",
   "square": "quadrat",
   "Start Date": "Data d'inici",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Color del text",
   "Text is required": "El text és obligatori",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "La sortida de text exclou el marc i la imatge.",
   "Text position": "Posició del text",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Hi ha massa codis QR per copiar al porta-retalls alhora. Si us plau, descarregueu-los com a SVG o PNG en lloc d'això.",
   "Tip: For best results, use a clear image with good lighting.": "Consell: Per obtenir els millors resultats, utilitza una imatge nítida amb bona il·luminació.",
   "Title": "Títol",
   "Toggle dark mode": "Canvia el mode fosc",
   "top": "dalt",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blocs complets)",
+  "Unicode (half-blocks)": "Unicode (meitats de blocs)",
   "Upload a CSV file": "Puja un fitxer CSV",
   "Upload image": "Pujar imatge",
   "Upload QR Code Image": "Pujar imatge del codi QR",
