+++ locales/cs.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Načteno z místního úložiště",
   "{count} piece(s) of data detected": "Zjištěno {count} kusů údajů",
   "{index} / {count} QR codes have been created.": "Byly vytvořeny QR kódy {index} / {count}.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7bitové znaky – fungují v prostém textu, e-mailech a terminálech.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7bitové, s obrácenou polaritou (znak # představuje pozadí) — funguje v prostém textu, e-mailech a terminálech.",
   "Add frame": "Přidat rám",
   "Address": "Adresa",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Zavřít dialogové okno",
   "Close guide": "Zavřít průvodce",
   "Close scanner": "Zavřít skener",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktní: dvě řady zhuštěné do jedné řádky pomocí znaků ▀ ▄ █.",
   "Configure frame settings": "Konfigurace nastavení rámu",
   "Contact Card": "Karta kontaktu",
-  "Copied": "Copied",
+  "Copied": "Zkopírováno",
   "Copy QR Code to clipboard": "Zkopírovat QR kód do schránky",
   "Copy to clipboard": "Kopírovat do schránky",
   "Corners Dot color": "Barva rohů tečky",
@@ -113,10 +113,10 @@
   "dots": "tečky",
   "Dots color": "Barva teček",
   "Dots type": "Typ teček",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Stáhnout soubor .md",
+  "Download .txt": "Stáhnout soubor .txt",
+  "Download all as .md": "Stáhnout vše jako .md",
+  "Download all as .txt": "Stáhnout vše jako soubor .txt",
   "Download QR Code as JPG": "Stáhnout QR kód jako JPG",
   "Download QR Code as PNG": "Stáhnout QR kód jako PNG",
   "Download QR Code as SVG": "Stáhnout QR kód jako SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Příklad souboru (vCard)",
   "Example with Custom Filenames": "Příklad s vlastními názvy souborů",
   "Export": "Exportovat",
-  "Export as text": "Export as text",
+  "Export as text": "Exportovat jako text",
   "Export QR code": "Exportovat QR kód",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportovat QR kód jako text ve formátu ASCII nebo Unicode",
   "extra-rounded": "krajně zakulacené",
   "Failed to copy to clipboard": "Kopírování do schránky se nezdařilo",
   "Failed to load changelog": "Nepodařilo se načíst seznam změn",
@@ -161,12 +161,12 @@
   "First Name": "Jméno",
   "First row preview:": "Náhled prvního řádku:",
   "Font family": "Rodina písem",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Štítek rámečku není zahrnut do exportu textu – výstup v kódování ASCII obsahuje pouze moduly QR.",
   "Frame preset": "Předvolba rámečku",
   "Frame settings": "Nastavení snímků",
   "Frame style": "Styl rámu",
   "Frame text": "Text rámečku",
-  "Generating": "Generating",
+  "Generating": "Generování",
   "GitHub repository for this project": "GitHub úložiště pro tento projekt",
   "Have a beautiful day!": "Přeji krásný den!",
   "Have nice day!": "Mějte příjemný den!",
@@ -204,7 +204,7 @@
   "Message": "Zpráva",
   "Mini QR Code Generator": "Mini generátor QR kódů",
   "MiniQR's QR rendering engine has been updated": "Byl aktualizován engine pro vykreslování QR kódů v aplikaci MiniQR",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Většina čteček QR kódů tento formát nerozpozná, protože znaky # vytvářejí mezery mezi řádky. Pro export v čitelné podobě použijte Unicode (poloviční/celé bloky).",
   "Must be between 0 and 1": "Musí být v rozmezí od 0 do 1",
   "Network Name (SSID)": "Název sítě (SSID)",
   "Network password (case-sensitive)": "Síťové heslo (rozlišuje velká a malá písmena)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Nejprve nahrajte soubor",
   "Position": "Pozice",
   "Preset": "Předvolba",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Náhled na první řádek",
   "Processing...": "Zpracovávám...",
   "QR Code Content": "Obsah QR kódu",
   "QR code preset": "Přednastavený QR kód",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Jednoduchá URL adresa/text",
   "Single export": "Jediný export",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Plné blokové znaky (██) – nejlepší věrnost zobrazení v moderních fontech.",
   "Sponsor": "Sponzor",
   "square": "čtverec",
   "Start Date": "Datum zahájení",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Barva textu",
   "Text is required": "Text je vyžadován",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Výstupní text nezahrnuje rámeček ani obrázek",
   "Text position": "Pozice textu",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Je příliš mnoho QR kódů ke zkopírování do schránky najednou. Stáhněte je prosím jako SVG nebo PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Tip: Pro nejlepší výsledky použijte čistý obrázek s dobrým osvětlením.",
   "Title": "Název",
   "Toggle dark mode": "Přepnout tmavý režim",
   "top": "nahoře",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (celé bloky)",
+  "Unicode (half-blocks)": "Unicode (polobloky)",
   "Upload a CSV file": "Nahrání souboru CSV",
   "Upload image": "Nahrát obrázek",
   "Upload QR Code Image": "Nahrát obrázek QR kódu",
