+++ locales/da.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Indlæst fra lokalt lager",
   "{count} piece(s) of data detected": "{count} styk detekteret data",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-koder er blevet oprettet.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bit-tegn — fungerer i almindelig tekst, e-mail og terminaler.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, omvendt polaritet (# er baggrunden) — fungerer i almindelig tekst, e-mail og terminaler.",
   "Add frame": "Tilføj ramme",
   "Address": "Adresse",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Luk dialogboksen",
   "Close guide": "Luk guide",
   "Close scanner": "Luk scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompakt: to rækker samlet i én linje ved hjælp af tegnene ▀ ▄ █.",
   "Configure frame settings": "Konfigurer rammeindstillinger",
   "Contact Card": "Kontaktkort",
-  "Copied": "Copied",
+  "Copied": "Kopieret",
   "Copy QR Code to clipboard": "Kopier QR-kode til udklipsholder",
   "Copy to clipboard": "Kopiér til udklipsholder",
   "Corners Dot color": "Hjørner Dot farve",
@@ -115,8 +115,8 @@
   "Dots type": "Prikker type",
   "Download .md": "Download .md",
   "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download all as .md": "Download det hele som .md",
+  "Download all as .txt": "Download det hele som .txt",
   "Download QR Code as JPG": "Download QR-kode som JPG",
   "Download QR Code as PNG": "Download QR-kode som PNG",
   "Download QR Code as SVG": "Download QR-kode som SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Eksempel på fil (vCard)",
   "Example with Custom Filenames": "Eksempel med brugerdefinerede filnavne",
   "Export": "Eksporter",
-  "Export as text": "Export as text",
+  "Export as text": "Eksporter som tekst",
   "Export QR code": "Eksporter QR-kode",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksporter QR-kode som ASCII- eller Unicode-tekst",
   "extra-rounded": "ekstra afrundet",
   "Failed to copy to clipboard": "Kunne ikke kopiere til udklipsholder",
   "Failed to load changelog": "Kunne ikke indlæse changelog",
@@ -161,12 +161,12 @@
   "First Name": "Fornavn",
   "First row preview:": "Forhåndsvisning af første række:",
   "Font family": "Skrifttypefamilie",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Rammeetiketten er udeladt fra teksteksporten — ASCII-udskriften indeholder kun QR-modulerne.",
   "Frame preset": "Forudindstillet ramme",
   "Frame settings": "Indstillinger for billeder",
   "Frame style": "Ramme stil",
   "Frame text": "Ramme-tekst",
-  "Generating": "Generating",
+  "Generating": "Generering",
   "GitHub repository for this project": "GitHub repository for dette projekt",
   "Have a beautiful day!": "Hav en dejlig dag!",
   "Have nice day!": "Hav en dejlig dag!",
@@ -204,7 +204,7 @@
   "Message": "Besked",
   "Mini QR Code Generator": "Mini QR-kodegenerator",
   "MiniQR's QR rendering engine has been updated": "MiniQRs QR-renderingmotor er blevet opdateret",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "De fleste QR-scannere kan ikke læse dette format, da #-tegnene efterlader mellemrum mellem rækkerne. Brug Unicode (halv-/helblokke) for at få en eksport, der kan scannes.",
   "Must be between 0 and 1": "Skal være mellem 0 og 1",
   "Network Name (SSID)": "Netværksnavn (SSID)",
   "Network password (case-sensitive)": "Netværksadgangskode (skelner mellem store og små bogstaver)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Upload først en fil",
   "Position": "Placering",
   "Preset": "Forudindstillet",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Forhåndsvisning af første række",
   "Processing...": "Behandler...",
   "QR Code Content": "Indhold af QR-kode",
   "QR code preset": "Forudindstillet QR-kode",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Enkel URL/tekst",
   "Single export": "Enkelt eksport",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Fyldte bloktegn (██) — giver den bedste gengivelse i moderne skrifttyper.",
   "Sponsor": "Sponsor",
   "square": "firkant",
   "Start Date": "Startdato",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Tekst farve",
   "Text is required": "Tekst er påkrævet",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Tekstudskriften omfatter ikke rammen og billedet",
   "Text position": "Tekst position",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Der er for mange QR-koder til at blive kopieret til udklipsholderen på én gang. Download dem som SVG eller PNG i stedet.",
   "Tip: For best results, use a clear image with good lighting.": "Tip: For de bedste resultater, brug et klart billede med god belysning.",
   "Title": "Titel",
   "Toggle dark mode": "Slå mørk tilstand til/fra",
   "top": "top",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (hele blokke)",
+  "Unicode (half-blocks)": "Unicode (halvblokke)",
   "Upload a CSV file": "Upload en CSV-fil",
   "Upload image": "Upload billede",
   "Upload QR Code Image": "Upload QR Kode Billede",
