+++ locales/de.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Zuletzt lokal gespeichert",
   "{count} piece(s) of data detected": "{count} Stück(e) der erkannten Daten",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-Codes wurden erstellt.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-Bit-Zeichen – funktionieren in reinem Text, E-Mails und Terminals.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-Bit, umgekehrte Polarität (# ist der Hintergrund) – funktioniert in Klartext, E-Mails und Terminals.",
   "Add frame": "Rahmen hinzufügen",
   "Address": "Adresse",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Dialog schließen",
   "Close guide": "Leitfaden schließen",
   "Close scanner": "Scanner schließen",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompakt: Zwei Reihen, die mithilfe der Zeichen ▀ ▄ █ in einer Zeile zusammengefasst sind.",
   "Configure frame settings": "Konfigurieren der Rahmeneinstellungen",
   "Contact Card": "Kontaktkarte",
-  "Copied": "Copied",
+  "Copied": "Kopiert",
   "Copy QR Code to clipboard": "QR-Code in die Zwischenablage kopieren",
   "Copy to clipboard": "In Zwischenablage kopieren",
   "Corners Dot color": "Farbe der Eckpunkte",
@@ -113,10 +113,10 @@
   "dots": "Punkte",
   "Dots color": "Punktefarbe",
   "Dots type": "Typ der Punkte",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".md herunterladen",
+  "Download .txt": "Als .txt herunterladen",
+  "Download all as .md": "Alles als .md herunterladen",
+  "Download all as .txt": "Alles als .txt herunterladen",
   "Download QR Code as JPG": "QR-Code als JPG herunterladen",
   "Download QR Code as PNG": "QR-Code als PNG herunterladen",
   "Download QR Code as SVG": "QR-Code als SVG herunterladen",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Beispieldatei (vCard)",
   "Example with Custom Filenames": "Beispiel mit benutzerdefinierten Dateinamen",
   "Export": "Exportieren",
-  "Export as text": "Export as text",
+  "Export as text": "Als Text exportieren",
   "Export QR code": "QR-Code exportieren",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR-Code als ASCII- oder Unicode-Text exportieren",
   "extra-rounded": "extra gerundet",
   "Failed to copy to clipboard": "Fehler beim Kopieren in die Zwischenablage",
   "Failed to load changelog": "Änderungsprotokoll konnte nicht geladen werden",
@@ -161,12 +161,12 @@
   "First Name": "Vorname",
   "First row preview:": "Vorschau der ersten Zeile:",
   "Font family": "Schriftfamilie",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Die Rahmenbeschriftung wird beim Textexport nicht berücksichtigt – die ASCII-Ausgabe enthält nur die QR-Module.",
   "Frame preset": "Rahmen-Voreinstellung",
   "Frame settings": "Rahmeneinstellungen",
   "Frame style": "Rahmenstil",
   "Frame text": "Rahmentext",
-  "Generating": "Generating",
+  "Generating": "Erstellen",
   "GitHub repository for this project": "GitHub-Repository für dieses Projekt",
   "Have a beautiful day!": "Ich wünsche Ihnen einen schönen Tag!",
   "Have nice day!": "Ich wünsche dir einen schönen Tag!",
@@ -204,7 +204,7 @@
   "Message": "Nachricht",
   "Mini QR Code Generator": "Mini QR-Code-Generator",
   "MiniQR's QR rendering engine has been updated": "Die QR-Rendering-Engine von MiniQR wurde aktualisiert",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Die meisten QR-Scanner können dieses Format nicht lesen, da die #-Zeichen Lücken zwischen den Zeilen hinterlassen. Verwenden Sie Unicode (Halb-/Vollblöcke), um einen scanbaren Export zu erhalten.",
   "Must be between 0 and 1": "Muss zwischen 0 und 1 liegen",
   "Network Name (SSID)": "Netzwerkname (SSID)",
   "Network password (case-sensitive)": "Netzwerkpasswort (Groß- und Kleinschreibung beachten)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Bitte laden Sie zuerst eine Datei hoch",
   "Position": "Position",
   "Preset": "Voreinstellung",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Vorschau der ersten Zeile",
   "Processing...": "Verarbeitung...",
   "QR Code Content": "QR-Code-Inhalt",
   "QR code preset": "QR-Code-Voreinstellung",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Einfache URL/Text",
   "Single export": "Einzelner Export",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Gefüllte Blockzeichen (██) – beste Wiedergabetreue bei modernen Schriftarten.",
   "Sponsor": "Sponsor",
   "square": "quadratisch",
   "Start Date": "Datum des Beginns",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Textfarbe",
   "Text is required": "Text ist erforderlich",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Die Textausgabe erfolgt ohne Rahmen und Bild",
   "Text position": "Textposition",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Es gibt zu viele QR-Codes, um auf einmal in die Zwischenablage kopiert werden zu können. Bitte laden Sie sie stattdessen als SVG oder PNG herunter.",
   "Tip: For best results, use a clear image with good lighting.": "Tipp: Verwenden Sie für die besten Ergebnisse ein klares Bild mit guter Beleuchtung.",
   "Title": "Titel",
   "Toggle dark mode": "Dunkler Modus umschalten",
   "top": "oben",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (Vollblöcke)",
+  "Unicode (half-blocks)": "Unicode (Halbblöcke)",
   "Upload a CSV file": "CSV-Datei hochladen",
   "Upload image": "Bild hochladen",
   "Upload QR Code Image": "QR-Code-Bild hochladen",
