+++ locales/el.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Φορτώθηκε από τοπική αποθήκευση",
   "{count} piece(s) of data detected": "{count} ανιχνευθέντα δεδομένα",
   "{index} / {count} QR codes have been created.": "Έχουν δημιουργηθεί {index} / {count} κωδικοί QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Χαρακτήρες 7-bit — λειτουργούν σε απλό κείμενο, ηλεκτρονικό ταχυδρομείο και τερματικά.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, αντίστροφη πολικότητα (το σύμβολο # είναι το φόντο) — λειτουργεί σε απλό κείμενο, email και τερματικά.",
   "Add frame": "Προσθήκη πλαισίου",
   "Address": "Διεύθυνση",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Κλείσιμο διαλόγου",
   "Close guide": "Κλείσιμο οδηγού",
   "Close scanner": "Κλείσιμο σαρωτή",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Συμπαγής: δύο σειρές συμπυκνωμένες σε μία γραμμή με τη χρήση των χαρακτήρων ▀ ▄ █.",
   "Configure frame settings": "Διαμόρφωση ρυθμίσεων πλαισίου",
   "Contact Card": "Κάρτα Επαφής",
-  "Copied": "Copied",
+  "Copied": "Αντιγραφή",
   "Copy QR Code to clipboard": "Αντιγραφή κώδικα QR στο πρόχειρο",
   "Copy to clipboard": "Αντιγραφή στο πρόχειρο",
   "Corners Dot color": "Γωνίες χρώμα Dot",
@@ -113,10 +113,10 @@
   "dots": "κουκίδες",
   "Dots color": "Χρώμα κουκκίδων",
   "Dots type": "Τύπος κουκκίδων",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Λήψη .md",
+  "Download .txt": "Λήψη αρχείου .txt",
+  "Download all as .md": "Λήψη όλων ως αρχεία .md",
+  "Download all as .txt": "Λήψη όλων ως αρχείο .txt",
   "Download QR Code as JPG": "Κατεβάστε τον κωδικό QR ως JPG",
   "Download QR Code as PNG": "Κατεβάστε τον κωδικό QR ως PNG",
   "Download QR Code as SVG": "Κατεβάστε τον κωδικό QR ως SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Παράδειγμα αρχείου (vCard)",
   "Example with Custom Filenames": "Παράδειγμα με προσαρμοσμένα ονόματα αρχείων",
   "Export": "Εξαγωγή",
-  "Export as text": "Export as text",
+  "Export as text": "Εξαγωγή ως κείμενο",
   "Export QR code": "Εξαγωγή κωδικού QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Εξαγωγή κωδικού QR ως κείμενο ASCII ή Unicode",
   "extra-rounded": "εξαιρετικά στρογγυλεμένος",
   "Failed to copy to clipboard": "Αποτυχία αντιγραφής στο πρόχειρο",
   "Failed to load changelog": "Απέτυχε να φορτώσει το changelog",
@@ -161,12 +161,12 @@
   "First Name": "Όνομα",
   "First row preview:": "Προεπισκόπηση πρώτης γραμμής:",
   "Font family": "Οικογένεια γραμματοσειρών",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Η ετικέτα πλαισίου δεν περιλαμβάνεται στην εξαγωγή κειμένου — η έξοδος ASCII περιέχει μόνο τις μονάδες QR.",
   "Frame preset": "Προεπιλογή πλαισίου",
   "Frame settings": "Ρυθμίσεις πλαισίου",
   "Frame style": "Στυλ πλαισίου",
   "Frame text": "Κείμενο πλαισίου",
-  "Generating": "Generating",
+  "Generating": "Δημιουργία",
   "GitHub repository for this project": "Αποθετήριο GitHub για αυτό το έργο",
   "Have a beautiful day!": "Να έχετε μια όμορφη μέρα!",
   "Have nice day!": "Καλή μέρα!",
@@ -204,7 +204,7 @@
   "Message": "Μήνυμα",
   "Mini QR Code Generator": "Μίνι γεννήτρια QR κωδικών",
   "MiniQR's QR rendering engine has been updated": "Ο μηχανισμός απόδοσης QR της MiniQR έχει ενημερωθεί",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Οι περισσότεροι σαρωτές QR δεν αναγνωρίζουν αυτή τη μορφή, καθώς τα σύμβολα # δημιουργούν κενά μεταξύ των σειρών. Χρησιμοποιήστε Unicode (μισά/πλήρη μπλοκ) για εξαγωγή που μπορεί να σαρωθεί.",
   "Must be between 0 and 1": "Πρέπει να είναι μεταξύ 0 και 1",
   "Network Name (SSID)": "Όνομα δικτύου (SSID)",
   "Network password (case-sensitive)": "Κωδικός πρόσβασης δικτύου (ευαίσθητη στην πεζότητα)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Παρακαλώ ανεβάστε ένα αρχείο πρώτα",
   "Position": "Θέση",
   "Preset": "Προεπιλογή",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Προεπισκόπηση της πρώτης σειράς",
   "Processing...": "Επεξεργασία...",
   "QR Code Content": "Περιεχόμενο QR Code",
   "QR code preset": "Προεπιλεγμένος κωδικός QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Απλή διεύθυνση URL/κείμενο",
   "Single export": "Ενιαία εξαγωγή",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Συμπαγείς χαρακτήρες (██) — η καλύτερη πιστότητα στις σύγχρονες γραμματοσειρές.",
   "Sponsor": "Χορηγός",
   "square": "τετράγωνο",
   "Start Date": "Ημερομηνία έναρξης",
@@ -291,15 +291,15 @@
   "Text": "Κείμενο",
   "Text color": "Χρώμα κειμένου",
   "Text is required": "Απαιτείται κείμενο",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Η έξοδος κειμένου δεν περιλαμβάνει το πλαίσιο και την εικόνα",
   "Text position": "Θέση κειμένου",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Υπάρχουν πάρα πολλοί κωδικοί QR για να αντιγραφούν στο πρόχειρο ταυτόχρονα. Παρακαλώ κατεβάστε τους ως SVG ή PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Συμβουλή: Για καλύτερα αποτελέσματα, χρησιμοποιήστε μια σαφή εικόνα με καλό φωτισμό.",
   "Title": "Τίτλος",
   "Toggle dark mode": "Εναλλαγή σκοτεινής λειτουργίας",
   "top": "κορυφή",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (πλήρη μπλοκ)",
+  "Unicode (half-blocks)": "Unicode (μισά τετράγωνα)",
   "Upload a CSV file": "Ανεβάστε ένα αρχείο CSV",
   "Upload image": "Μεταφόρτωση εικόνας",
   "Upload QR Code Image": "Μεταφόρτωση Εικόνας Κωδικού Qr",
