+++ locales/et.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Laetud kohalikust ladustusest",
   "{count} piece(s) of data detected": "{count} tuvastatud andmed",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-koodid on loodud.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bitised tähemärgid — sobivad kasutamiseks lihttekstis, e-kirjades ja terminalides.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bitine, pööratud polaarsus (sümbol # tähistab tausta) — toimib tavalises tekstis, e-kirjades ja terminalides.",
   "Add frame": "Lisa raam",
   "Address": "Aadress",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Dialoogi sulgemine",
   "Close guide": "Sulge juhend",
   "Close scanner": "Sulge skanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktne: kaks rida on ühte ritta koondatud, kasutades sümboleid ▀ ▄ █.",
   "Configure frame settings": "Kaadri seadete konfigureerimine",
   "Contact Card": "Kontaktkaart",
-  "Copied": "Copied",
+  "Copied": "Kopeeritud",
   "Copy QR Code to clipboard": "Kopeeri QR-kood lõikelauale",
   "Copy to clipboard": "Kopeeri lõikelauda",
   "Corners Dot color": "Nurgad Punkti värvus",
@@ -113,10 +113,10 @@
   "dots": "punktid",
   "Dots color": "Punkti värvus",
   "Dots type": "Punktide tüüp",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Lae alla .md",
+  "Download .txt": "Lae alla .txt",
+  "Download all as .md": "Lae kõik alla .md-vormingus",
+  "Download all as .txt": "Lae kõik alla .txt-failina",
   "Download QR Code as JPG": "Lae QR-kood alla kui JPG",
   "Download QR Code as PNG": "QR-koodi allalaadimine PNG-formaadis",
   "Download QR Code as SVG": "Laadige QR-kood alla SVG-formaadis",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Näidisfail (vCard)",
   "Example with Custom Filenames": "Näide kohandatud failinimedega",
   "Export": "Eksport",
-  "Export as text": "Export as text",
+  "Export as text": "Ekspordi tekstina",
   "Export QR code": "Ekspordi QR-kood",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Ekspordi QR-kood ASCII- või Unicode-tekstina",
   "extra-rounded": "ekstra ümarad",
   "Failed to copy to clipboard": "Kopeerimine lõikelauale ebaõnnestus",
   "Failed to load changelog": "Ei õnnestunud laadida muutuste logi",
@@ -161,12 +161,12 @@
   "First Name": "Eesnimi",
   "First row preview:": "Esimese rea eelvaade:",
   "Font family": "Kirjasort",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Raami silt jäetakse teksti ekspordist välja – ASCII-vormingus väljund sisaldab ainult QR-mooduleid.",
   "Frame preset": "Raami eelseadistus",
   "Frame settings": "Raami seaded",
   "Frame style": "Raami stiil",
   "Frame text": "Raami tekst",
-  "Generating": "Generating",
+  "Generating": "Loomine",
   "GitHub repository for this project": "Selle projekti GitHubi repositoorium",
   "Have a beautiful day!": "Ilusat päeva!",
   "Have nice day!": "Ilusat päeva!",
@@ -204,7 +204,7 @@
   "Message": "Sõnum",
   "Mini QR Code Generator": "Mini QR-koodi generaator",
   "MiniQR's QR rendering engine has been updated": "MiniQR-i QR-koodide renderdamise mootor on uuendatud",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Enamik QR-skannereid ei suuda seda formaati lugeda, kuna sümbolid # jätavad ridade vahele tühikuid. Kasutage skannitava ekspordi jaoks Unicode'i (pool- või täisplokid).",
   "Must be between 0 and 1": "Peab olema vahemikus 0–1",
   "Network Name (SSID)": "Võrgu nimi (SSID)",
   "Network password (case-sensitive)": "Võrguparool (sõltuvalt suur- ja väiketähendusest)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Palun laadige kõigepealt fail üles",
   "Position": "Positsioon",
   "Preset": "Eelseadistatud",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Esimese rea eelvaade",
   "Processing...": "Töötlemine...",
   "QR Code Content": "QR-koodi sisu",
   "QR code preset": "QR-koodi eelseadistus",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Lihtne URL/tekst",
   "Single export": "Ühekordne eksport",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Täisplokk-märgid (██) — parim täpsus tänapäevastes fondides.",
   "Sponsor": "sponsor",
   "square": "ruudukujuline",
   "Start Date": "Alguskuupäev",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Teksti värvus",
   "Text is required": "Tekst on nõutav",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Tekstiväljund ei sisalda raami ega pilti",
   "Text position": "Teksti asukoht",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "QR-koode on liiga palju, et neid korraga lõikelauale kopeerida. Palun laadige need hoopis SVG või PNG kujul alla.",
   "Tip: For best results, use a clear image with good lighting.": "Vihje: Parima tulemuse saamiseks kasutage selge ja hea valgustusega pilti.",
   "Title": "Pealkiri",
   "Toggle dark mode": "Lülita tume režiim sisse",
   "top": "tipp",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (täisplokid)",
+  "Unicode (half-blocks)": "Unicode (poolplokid)",
   "Upload a CSV file": "CSV-faili üleslaadimine",
   "Upload image": "Pildi üleslaadimine",
   "Upload QR Code Image": "QR-koodi pildi üleslaadimine",
