+++ locales/fa.json
@@ -51,11 +51,11 @@
   "Last saved locally": "بارگذاری شده از ذخیره‌سازی محلی",
   "{count} piece(s) of data detected": "{count} تکه‌های دادهٔ شناسایی‌شده",
   "{index} / {count} QR codes have been created.": "{index} / {count} کدهای QR ایجاد شده‌اند.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "کاراکترهای ۷ بیتی — در متن ساده، ایمیل و ترمینال‌ها کار می‌کند.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "۷ بیت، قطبیت معکوس (# پس‌زمینه است) — در متن ساده، ایمیل و ترمینال‌ها کار می‌کند.",
   "Add frame": "افزودن قاب",
   "Address": "آدرس",
-  "ASCII": "ASCII",
+  "ASCII": "آیسای",
   "Available Fields": "میدان‌های موجود",
   "Background color": "رنگ پس‌زمینه",
   "Batch export": "صادرات دسته‌ای",
@@ -82,10 +82,10 @@
   "Close dialog": "پنجرهٔ گفتگو را ببندید",
   "Close guide": "بستن راهنما",
   "Close scanner": "بستن اسکنر",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "کوتاه: دو خط در یک خط فشرده شده با استفاده از کاراکترهای ▀ ▄ █.",
   "Configure frame settings": "تنظیمات فریم را پیکربندی کنید",
   "Contact Card": "کارت تماس",
-  "Copied": "Copied",
+  "Copied": "کپی شد",
   "Copy QR Code to clipboard": "کد QR را به کلیپ‌بورد کپی کنید",
   "Copy to clipboard": "کپی به کلیپ‌بورد",
   "Corners Dot color": "رنگ نقطه‌ای گوشه‌ها",
@@ -113,10 +113,10 @@
   "dots": "نقطه‌ها",
   "Dots color": "رنگ نقاط",
   "Dots type": "نوع نقاط",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "دانلود .md",
+  "Download .txt": "دانلود .txt",
+  "Download all as .md": "همه را به صورت .md دانلود کنید",
+  "Download all as .txt": "همه را به صورت .txt دانلود کنید",
   "Download QR Code as JPG": "دانلود کد QR به صورت JPG",
   "Download QR Code as PNG": "دانلود کد QR به صورت PNG",
   "Download QR Code as SVG": "دانلود کد QR به صورت SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "فایل نمونه (vCard)",
   "Example with Custom Filenames": "مثال با نام‌های فایل سفارشی",
   "Export": "صادرات",
-  "Export as text": "Export as text",
+  "Export as text": "صادرات به صورت متن",
   "Export QR code": "صادرات کد QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "کد QR را به صورت متن ASCII یا یونیکد صادر کنید",
   "extra-rounded": "بسیار گرد",
   "Failed to copy to clipboard": "کپی به کلیپ‌بورد با موفقیت انجام نشد",
   "Failed to load changelog": "بارگذاری لاگ تغییرات ناموفق بود.",
@@ -161,12 +161,12 @@
   "First Name": "نام",
   "First row preview:": "پیش‌نمایش ردیف اول:",
   "Font family": "خانواده قلم",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "برچسب قاب در خروجی متن لحاظ نمی‌شود — خروجی ASCII تنها شامل ماژول‌های QR است.",
   "Frame preset": "پیش‌تنظیم قاب",
   "Frame settings": "تنظیمات قاب",
   "Frame style": "سبک قاب",
   "Frame text": "متن قاب",
-  "Generating": "Generating",
+  "Generating": "تولید",
   "GitHub repository for this project": "انبار گیت‌هاب برای این پروژه",
   "Have a beautiful day!": "روز زیبایی داشته باشید!",
   "Have nice day!": "روز خوبی داشته باشید!",
@@ -204,7 +204,7 @@
   "Message": "پیام",
   "Mini QR Code Generator": "مینی تولیدکننده کد QR",
   "MiniQR's QR rendering engine has been updated": "موتور رندرینگ QR مینی‌کیوآر به‌روزرسانی شده است.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "اکثر اسکنرهای QR این فرمت را نمی‌خوانند زیرا نمادهای # بین سطرها فاصله ایجاد می‌کنند. برای خروجی قابل اسکن از یونیکد (بلوک‌های نیم‌بلوکی/تمام‌بلوکی) استفاده کنید.",
   "Must be between 0 and 1": "باید بین ۰ و ۱ باشد",
   "Network Name (SSID)": "نام شبکه (SSID)",
   "Network password (case-sensitive)": "رمز عبور شبکه (حساس به حروف بزرگ و کوچک)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "لطفاً ابتدا یک فایل را آپلود کنید.",
   "Position": "موقعیت",
   "Preset": "پیش‌تنظیم",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "پیش‌نمایش ردیف اول",
   "Processing...": "در حال پردازش...",
   "QR Code Content": "محتوای کد QR",
   "QR code preset": "پیش‌تنظیم کد QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "پیوند ساده/متن",
   "Single export": "صادرات تکی",
   "SMS": "پیامک",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "کاراکترهای بلوکی توپر (██) — بهترین وضوح در فونت‌های مدرن.",
   "Sponsor": "حامی",
   "square": "مربع",
   "Start Date": "تاریخ شروع",
@@ -291,15 +291,15 @@
   "Text": "متن",
   "Text color": "رنگ متن",
   "Text is required": "متن الزامی است",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "خروجی متن، قاب و تصویر را شامل نمی‌شود.",
   "Text position": "موقعیت متن",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "تعداد کدهای QR برای کپی هم‌زمان به کلیپ‌بورد بیش از حد است. لطفاً آن‌ها را به‌صورت SVG یا PNG دانلود کنید.",
   "Tip: For best results, use a clear image with good lighting.": "نکته: برای بهترین نتیجه، از تصویری واضح با نورپردازی مناسب استفاده کنید.",
   "Title": "عنوان",
   "Toggle dark mode": "مُد تاریک",
   "top": "بالا",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "یونیکد (بلوک‌های کامل)",
+  "Unicode (half-blocks)": "یونیکد (نیم‌بلاک‌ها)",
   "Upload a CSV file": "یک فایل CSV را بارگذاری کنید",
   "Upload image": "بارگذاری تصویر",
   "Upload QR Code Image": "بارگذاری تصویر کد QR",
