+++ locales/fi.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Ladattu paikallisesta tallennustilasta",
   "{count} piece(s) of data detected": "{count} havaitut tiedot",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-koodit on luotu.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bittiset merkit — toimivat tekstimuodossa, sähköpostissa ja päätelaitteissa.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bittinen, käänteinen napaisuus (# on tausta) — toimii tekstimuodossa, sähköpostissa ja päätelaitteissa.",
   "Add frame": "Lisää kehys",
   "Address": "Osoite",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Sulje valintaikkuna",
   "Close guide": "Sulje opas",
   "Close scanner": "Sulje skanneri",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Tiivis: kaksi riviä tiivistettynä yhdeksi riviksi ▀ ▄ █-merkkien avulla.",
   "Configure frame settings": "Kehysasetusten määrittäminen",
   "Contact Card": "Yhteystietokortti",
-  "Copied": "Copied",
+  "Copied": "Kopioitu",
   "Copy QR Code to clipboard": "Kopioi QR-koodi leikepöydälle",
   "Copy to clipboard": "Kopioi leikepöydälle",
   "Corners Dot color": "Kulmat Pisteiden väri",
@@ -113,10 +113,10 @@
   "dots": "pistettä",
   "Dots color": "Pisteiden väri",
   "Dots type": "Pisteiden tyyppi",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Lataa .md",
+  "Download .txt": "Lataa .txt-tiedosto",
+  "Download all as .md": "Lataa kaikki .md-tiedostona",
+  "Download all as .txt": "Lataa kaikki .txt-tiedostona",
   "Download QR Code as JPG": "Lataa QR-koodi JPG",
   "Download QR Code as PNG": "Lataa QR-koodi PNG",
   "Download QR Code as SVG": "Lataa QR-koodi SVG:nä",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Esimerkkitiedosto (vCard)",
   "Example with Custom Filenames": "Esimerkki mukautetuilla tiedostonimillä",
   "Export": "Vie",
-  "Export as text": "Export as text",
+  "Export as text": "Vie tekstinä",
   "Export QR code": "Vie QR-koodi",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Vie QR-koodi ASCII- tai Unicode-tekstinä",
   "extra-rounded": "ekstrapyöristetty",
   "Failed to copy to clipboard": "Kopiointi leikepöydälle epäonnistui",
   "Failed to load changelog": "Changelogin lataaminen epäonnistui",
@@ -161,12 +161,12 @@
   "First Name": "Etunimi",
   "First row preview:": "Ensimmäisen rivin esikatselu:",
   "Font family": "Fonttiperhe",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Kehysmerkintä jätetään pois tekstivientiä varten — ASCII-tuloste sisältää vain QR-moduulit.",
   "Frame preset": "Kehyksen esiasetus",
   "Frame settings": "Kehyksen asetukset",
   "Frame style": "Kehyksen tyyli",
   "Frame text": "Kehyksen teksti",
-  "Generating": "Generating",
+  "Generating": "Luominen",
   "GitHub repository for this project": "GitHub repo tälle projektille",
   "Have a beautiful day!": "Kaunista päivää!",
   "Have nice day!": "Pidä mukavaa päivää!",
@@ -204,7 +204,7 @@
   "Message": "Viesti",
   "Mini QR Code Generator": "Pieni QR-koodigeneraattori",
   "MiniQR's QR rendering engine has been updated": "MiniQR:n QR-koodien renderointimoottori on päivitetty",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Useimmat QR-lukijat eivät tunnista tätä muotoa, koska #-merkit jättävät rivejen väliin tyhjiä kohtia. Käytä Unicodea (puoli- tai täysmerkkijonoja), jotta tiedosto voidaan lukea skannerilla.",
   "Must be between 0 and 1": "Arvon on oltava välillä 0–1",
   "Network Name (SSID)": "Verkon nimi (SSID)",
   "Network password (case-sensitive)": "Verkon salasana (isojen ja pienten kirjainten erittely)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Ole hyvä ja lataa tiedosto ensin",
   "Position": "Sijainti",
   "Preset": "Esiasetus",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Esikatselu ensimmäisestä rivistä",
   "Processing...": "Käsitellään...",
   "QR Code Content": "QR-koodin sisältö",
   "QR code preset": "QR-koodin esiasetus",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Yksinkertainen URL/teksti",
   "Single export": "Yksittäinen vienti",
   "SMS": "tekstiviesti",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Yhtenäiset neliömerkit (██) — paras tarkkuus nykyaikaisissa fonteissa.",
   "Sponsor": "Sponsori",
   "square": "neliö",
   "Start Date": "Aloituspäivä",
@@ -291,15 +291,15 @@
   "Text": "Teksti",
   "Text color": "Tekstin väri",
   "Text is required": "Teksti vaaditaan",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Tekstin tulostuksessa kehystä ja kuvaa ei oteta huomioon",
   "Text position": "Tekstin sijainti",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "QR-koodeja on liian paljon, jotta ne voidaan kopioida leikepöydälle kerralla. Ole hyvä ja lataa ne sen sijaan SVG- tai PNG-muodossa.",
   "Tip: For best results, use a clear image with good lighting.": "Vinkki: Saat parhaat tulokset, käytä selkeää kuvaa hyvällä valaistuksella.",
   "Title": "Otsikko",
   "Toggle dark mode": "Vaihda tumma tila",
   "top": "ylhäällä",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (kokonaiset lohkot)",
+  "Unicode (half-blocks)": "Unicode (puolilohkot)",
   "Upload a CSV file": "Lataa CSV-tiedosto",
   "Upload image": "Lataa kuva",
   "Upload QR Code Image": "Lataa QR-koodikuva",
