+++ locales/fr.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Chargé depuis le stockage local",
   "{count} piece(s) of data detected": "{count} pièce(s) de données détectée(s)",
   "{index} / {count} QR codes have been created.": "Les codes QR {index} / {count} ont été créés.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Caractères 7 bits — fonctionne dans le texte brut, les e-mails et les terminaux.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bits, polarité inversée (le signe # correspond à l'arrière-plan) — fonctionne dans le texte brut, les e-mails et les terminaux.",
   "Add frame": "Ajouter un cadre",
   "Address": "Adresse",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Fermer le dialogue",
   "Close guide": "Fermer le guide",
   "Close scanner": "Fermer le scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact : deux lignes regroupées en une seule à l'aide des caractères ▀ ▄ █.",
   "Configure frame settings": "Configurer les paramètres du cadre",
   "Contact Card": "Carte de contact",
-  "Copied": "Copied",
+  "Copied": "Copié",
   "Copy QR Code to clipboard": "Copier le QR Code dans le presse-papiers",
   "Copy to clipboard": "Copier dans le presse-papiers",
   "Corners Dot color": "Couleur du point d'angle",
@@ -113,10 +113,10 @@
   "dots": "points",
   "Dots color": "Couleur des points",
   "Dots type": "Type de point",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Télécharger le fichier .md",
+  "Download .txt": "Télécharger le fichier .txt",
+  "Download all as .md": "Tout télécharger au format .md",
+  "Download all as .txt": "Tout télécharger au format .txt",
   "Download QR Code as JPG": "Télécharger le QR Code en JPG",
   "Download QR Code as PNG": "Télécharger le QR Code en PNG",
   "Download QR Code as SVG": "Télécharger le QR Code au format SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Fichier d'exemple (vCard)",
   "Example with Custom Filenames": "Exemple avec des noms de fichiers personnalisés",
   "Export": "Exportation",
-  "Export as text": "Export as text",
+  "Export as text": "Exporter au format texte",
   "Export QR code": "Exporter le code QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exporter le code QR au format texte ASCII ou Unicode",
   "extra-rounded": "extra-arrondi",
   "Failed to copy to clipboard": "Échec de la copie dans le presse-papiers",
   "Failed to load changelog": "Échec du chargement du changelog",
@@ -161,12 +161,12 @@
   "First Name": "Prénom",
   "First row preview:": "Prévisualisation de la première ligne:",
   "Font family": "Famille de polices",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "L'étiquette de trame est exclue de l'exportation de texte : le fichier ASCII généré ne contient que les modules QR.",
   "Frame preset": "Cadre prédéfini",
   "Frame settings": "Paramètres du cadre",
   "Frame style": "Style de cadre",
   "Frame text": "Texte du cadre",
-  "Generating": "Generating",
+  "Generating": "Génération",
   "GitHub repository for this project": "Dépôt GitHub pour ce projet",
   "Have a beautiful day!": "Bonne journée !",
   "Have nice day!": "Bonne journée!",
@@ -204,7 +204,7 @@
   "Message": "Message",
   "Mini QR Code Generator": "Générateur de mini codes QR",
   "MiniQR's QR rendering engine has been updated": "Le moteur de rendu QR de MiniQR a été mis à jour",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "La plupart des lecteurs de QR codes ne parviennent pas à lire ce format, car les caractères # laissent des espaces entre les lignes. Utilisez l'Unicode (blocs demi-largeur/pleine largeur) pour obtenir un fichier exportable lisible par les lecteurs de QR codes.",
   "Must be between 0 and 1": "Doit être compris entre 0 et 1",
   "Network Name (SSID)": "Nom du réseau (SSID)",
   "Network password (case-sensitive)": "Mot de passe du réseau (sensible à la casse)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Veuillez d'abord télécharger un fichier",
   "Position": "Position",
   "Preset": "Préréglé",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Aperçu de la première ligne",
   "Processing...": "Traitement en cours...",
   "QR Code Content": "Contenu du code QR",
   "QR code preset": "Code QR prédéfini",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL simple/Texte",
   "Single export": "Exportation simple",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Caractères en blocs pleins (██) — la meilleure fidélité avec les polices modernes.",
   "Sponsor": "Commanditaire",
   "square": "carré",
   "Start Date": "Date de début",
@@ -291,15 +291,15 @@
   "Text": "Texte",
   "Text color": "Couleur du texte",
   "Text is required": "Le texte est obligatoire",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Le texte affiché ne comprend ni le cadre ni l'image",
   "Text position": "Position du texte",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Il y a trop de codes QR pour être copié dans le presse-papiers à la fois. Veuillez les télécharger au format SVG ou PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Astuce : Pour de meilleurs résultats, utilisez une image claire avec un bon éclairage.",
   "Title": "Titre",
   "Toggle dark mode": "Activer/désactiver le mode sombre",
   "top": "en haut",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blocs complets)",
+  "Unicode (half-blocks)": "Unicode (demi-blocs)",
   "Upload a CSV file": "Télécharger un fichier CSV",
   "Upload image": "Charger une image",
   "Upload QR Code Image": "Télécharger l'image du QR Code",
