+++ locales/he.json
@@ -51,8 +51,8 @@
   "Last saved locally": "נטען מאחסון מקומי",
   "{count} piece(s) of data detected": "{count} נתון/נתונים שזוהו",
   "{index} / {count} QR codes have been created.": "{index} / {count} נוצרו קודי QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "תווים של 7 סיביות — מתאימים לטקסט רגיל, דואר אלקטרוני ומסופים.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 סיביות, קוטביות הפוכה (# הוא הרקע) — מתאים לטקסט רגיל, דואר אלקטרוני ומסופים.",
   "Add frame": "הוסף מסגרת",
   "Address": "כתובת",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "סגור דיאלוג",
   "Close guide": "סגור מדריך",
   "Close scanner": "סגור סורק",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "קומפקטי: שתי שורות דחוסות בשורה אחת באמצעות התווים ▀ ▄ █.",
   "Configure frame settings": "הגדר את הגדרות המסגרת",
   "Contact Card": "כרטיס יצירת קשר",
-  "Copied": "Copied",
+  "Copied": "הועתק",
   "Copy QR Code to clipboard": "העתק את קוד ה-QR ללוח",
   "Copy to clipboard": "העתק ללוח",
   "Corners Dot color": "פינות צבע נקודה",
@@ -113,10 +113,10 @@
   "dots": "נקודות",
   "Dots color": "צבע הנקודות",
   "Dots type": "סוג נקודות",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "הורד קובץ .md",
+  "Download .txt": "הורד קובץ .txt",
+  "Download all as .md": "הורד הכל כקובץ .md",
+  "Download all as .txt": "הורד את הכל כקובץ .txt",
   "Download QR Code as JPG": "הורד קוד QR כקובץ JPG",
   "Download QR Code as PNG": "הורד קוד QR כקובץ PNG",
   "Download QR Code as SVG": "הורד קוד QR כקובץ SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "קובץ לדוגמה (vCard)",
   "Example with Custom Filenames": "דוגמה עם שמות קבצים מותאמים אישית",
   "Export": "ייצוא",
-  "Export as text": "Export as text",
+  "Export as text": "ייצא כטקסט",
   "Export QR code": "ייצא קוד QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "ייצא קוד QR כטקסט ASCII או Unicode",
   "extra-rounded": "מעוגל במיוחד",
   "Failed to copy to clipboard": "לא ניתן להעתיק ללוח",
   "Failed to load changelog": "לא ניתן לטעון את יומן השינויים",
@@ -161,12 +161,12 @@
   "First Name": "שם פרטי",
   "First row preview:": "תצוגה מקדימה של השורה הראשונה:",
   "Font family": "משפחת גופנים",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "תווית המסגרת אינה נכללת בייצוא הטקסט — הפלט ב-ASCII מכיל את מודולי ה-QR בלבד.",
   "Frame preset": "הגדרת מסגרת קבועה מראש",
   "Frame settings": "הגדרות מסגרת",
   "Frame style": "סגנון מסגרת",
   "Frame text": "טקסט מסגרת",
-  "Generating": "Generating",
+  "Generating": "יצירה",
   "GitHub repository for this project": "מאגר GitHub עבור פרויקט זה",
   "Have a beautiful day!": "שיהיה לך יום נפלא!",
   "Have nice day!": "שיהיה לך יום נעים!",
@@ -204,7 +204,7 @@
   "Message": "הודעה",
   "Mini QR Code Generator": "מיני מחולל קוד QR",
   "MiniQR's QR rendering engine has been updated": "מנוע העיבוד של MiniQR עודכן",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "רוב סורקי ה-QR לא יקראו פורמט זה, מכיוון שהסימנים # יוצרים רווחים בין השורות. השתמשו ב-Unicode (בלוקים חצויים/מלאים) כדי לייצא קובץ שניתן לסרוק.",
   "Must be between 0 and 1": "חייב להיות בין 0 ל-1",
   "Network Name (SSID)": "שם הרשת (SSID)",
   "Network password (case-sensitive)": "סיסמת רשת (רגישה לאותיות רישיות)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "אנא העלה קובץ תחילה",
   "Position": "מיקום",
   "Preset": "הגדרה קבועה מראש",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "תצוגה מקדימה של השורה הראשונה",
   "Processing...": "עיבוד...",
   "QR Code Content": "תוכן קוד QR",
   "QR code preset": "קוד QR מוגדר מראש",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/טקסט פשוט",
   "Single export": "ייצוא יחיד",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "תווים ריקים (██) — רמת הדיוק הטובה ביותר בגופנים מודרניים.",
   "Sponsor": "נותן חסות",
   "square": "ריבוע",
   "Start Date": "תאריך התחלה",
@@ -291,15 +291,15 @@
   "Text": "טקסט",
   "Text color": "צבע הטקסט",
   "Text is required": "טקסט נדרש",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "הטקסט המוצג אינו כולל את המסגרת והתמונה",
   "Text position": "מיקום הטקסט",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "יש יותר מדי קודי QR מכדי להעתיק אותם ללוח בבת אחת. אנא הורד אותם כקובצי SVG או PNG.",
   "Tip: For best results, use a clear image with good lighting.": "טיפ: לקבלת תוצאות מיטביות, השתמש בתמונה ברורה עם תאורה טובה.",
   "Title": "כותרת",
   "Toggle dark mode": "הפעל מצב כהה",
   "top": "למעלה",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (בלוקים מלאים)",
+  "Unicode (half-blocks)": "יוניקוד (חצאי בלוקים)",
   "Upload a CSV file": "העלה קובץ CSV",
   "Upload image": "העלה תמונה",
   "Upload QR Code Image": "העלה תמונת קוד QR",
