+++ locales/hi.json
@@ -51,11 +51,11 @@
   "Last saved locally": "स्थानीय भंडारण से लोड किया गया",
   "{count} piece(s) of data detected": "{count} डेटा के हिस्से का पता चला",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR कोड बनाए गए हैं।",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-बिट वर्ण — सादा पाठ, ईमेल और टर्मिनलों में काम करता है।",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-बिट, उल्टी ध्रुवता (# पृष्ठभूमि है) — सादा पाठ, ईमेल और टर्मिनलों में काम करता है।",
   "Add frame": "फ़्रेम जोड़ें",
   "Address": "पता",
-  "ASCII": "ASCII",
+  "ASCII": "एएससीआईआई",
   "Available Fields": "उपलब्ध क्षेत्र",
   "Background color": "पृष्ठभूमि का रंग",
   "Batch export": "बैच निर्यात",
@@ -82,10 +82,10 @@
   "Close dialog": "संवाद बॉक्स बंद करें",
   "Close guide": "गाइड बंद करें",
   "Close scanner": "स्कैनर बंद करें",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "कॉम्पैक्ट: ▀ ▄ █ वर्णों का उपयोग करके दो पंक्तियों को एक ही पंक्ति में संकुचित किया गया।",
   "Configure frame settings": "फ़्रेम सेटिंग्स कॉन्फ़िगर करें",
   "Contact Card": "संपर्क कार्ड",
-  "Copied": "Copied",
+  "Copied": "नकल किया गया",
   "Copy QR Code to clipboard": "क्यूआर कोड को क्लिपबोर्ड पर कॉपी करें",
   "Copy to clipboard": "क्लिपबोर्ड में कॉपी करें",
   "Corners Dot color": "कोने डॉट रंग",
@@ -113,10 +113,10 @@
   "dots": "बिंदु",
   "Dots color": "बिंदुओं का रंग",
   "Dots type": "बिंदु प्रकार",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "डाउनलोड .md",
+  "Download .txt": "डाउनलोड .txt",
+  "Download all as .md": "सभी को .md के रूप में डाउनलोड करें",
+  "Download all as .txt": "सभी को .txt के रूप में डाउनलोड करें",
   "Download QR Code as JPG": "QR कोड को JPG के रूप में डाउनलोड करें",
   "Download QR Code as PNG": "QR कोड को PNG के रूप में डाउनलोड करें",
   "Download QR Code as SVG": "QR कोड को SVG के रूप में डाउनलोड करें",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "उदाहरण फ़ाइल (vCard)",
   "Example with Custom Filenames": "कस्टम फ़ाइलनामों के साथ उदाहरण",
   "Export": "निर्यात",
-  "Export as text": "Export as text",
+  "Export as text": "टेक्स्ट के रूप में निर्यात करें",
   "Export QR code": "क्यूआर कोड निर्यात करें",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR कोड को ASCII या यूनिकोड टेक्स्ट के रूप में निर्यात करें",
   "extra-rounded": "अति-गोल",
   "Failed to copy to clipboard": "क्लिपबोर्ड में कॉपी करने में असफल",
   "Failed to load changelog": "चेंजलॉग लोड करने में असफल रहा",
@@ -161,12 +161,12 @@
   "First Name": "पहला नाम",
   "First row preview:": "पहली पंक्ति पूर्वावलोकन:",
   "Font family": "फ़ॉन्ट परिवार",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "फ़्रेम लेबल टेक्स्ट एक्सपोर्ट से बाहर रखा गया है — ASCII आउटपुट में केवल QR मॉड्यूल होते हैं।",
   "Frame preset": "फ़्रेम प्रीसेट",
   "Frame settings": "फ़्रेम सेटिंग्स",
   "Frame style": "फ़्रेम शैली",
   "Frame text": "फ़्रेम टेक्स्ट",
-  "Generating": "Generating",
+  "Generating": "उत्पादित कर रहा है",
   "GitHub repository for this project": "इस परियोजना के लिए GitHub रिपॉजिटरी",
   "Have a beautiful day!": "आपका दिन मंगलमय हो!",
   "Have nice day!": "आपका दिन शुभ हो!",
@@ -204,7 +204,7 @@
   "Message": "संदेश",
   "Mini QR Code Generator": "मिनी क्यूआर कोड जनरेटर",
   "MiniQR's QR rendering engine has been updated": "मिनीक्यूआर का क्यूआर रेंडरिंग इंजन अपडेट किया गया है।",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "अधिकांश QR स्कैनर इस प्रारूप को नहीं पढ़ पाएंगे क्योंकि # ग्लिफ्स पंक्तियों के बीच रिक्त स्थान छोड़ देते हैं। स्कैन करने योग्य निर्यात के लिए यूनिकोड (अर्ध/पूर्ण ब्लॉक) का उपयोग करें।",
   "Must be between 0 and 1": "0 और 1 के बीच होना चाहिए",
   "Network Name (SSID)": "नेटवर्क का नाम (एसएसआईडी)",
   "Network password (case-sensitive)": "नेटवर्क पासवर्ड (अक्षर-संवेदनशील)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "कृपया पहले एक फ़ाइल अपलोड करें।",
   "Position": "पद",
   "Preset": "पूर्वनिर्धारित",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "पहली पंक्ति का पूर्वावलोकन",
   "Processing...": "प्रसंस्करण...",
   "QR Code Content": "क्यूआर कोड सामग्री",
   "QR code preset": "क्यूआर कोड पूर्वनिर्धारित",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "सरल यूआरएल/टेक्स्ट",
   "Single export": "एकल निर्यात",
   "SMS": "एसएमएस",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "ठोस ब्लॉक वर्ण (██) — आधुनिक फ़ॉन्ट्स में सर्वोत्तम स्पष्टता।",
   "Sponsor": "प्रायोजक",
   "square": "वर्ग",
   "Start Date": "आरंभ तिथि",
@@ -291,15 +291,15 @@
   "Text": "पाठ",
   "Text color": "पाठ का रंग",
   "Text is required": "टेक्स्ट आवश्यक है",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "टेक्स्ट आउटपुट में फ्रेम और छवि शामिल नहीं हैं।",
   "Text position": "पाठ की स्थिति",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "एक साथ क्लिपबोर्ड पर कॉपी करने के लिए बहुत सारे QR कोड हैं। कृपया इन्हें इसके बजाय SVG या PNG के रूप में डाउनलोड करें।",
   "Tip: For best results, use a clear image with good lighting.": "सुझाव: सर्वोत्तम परिणामों के लिए, अच्छी रोशनी वाली एक स्पष्ट छवि का उपयोग करें।",
   "Title": "शीर्षक",
   "Toggle dark mode": "डार्क मोड टॉगल करें",
   "top": "शीर्ष",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "यूनिकोड (पूर्ण-ब्लॉक)",
+  "Unicode (half-blocks)": "यूनिकोड (अर्ध-ब्लॉक)",
   "Upload a CSV file": "CSV फ़ाइल अपलोड करें",
   "Upload image": "छवि अपलोड करें",
   "Upload QR Code Image": "क्यूआर कोड छवि अपलोड करें",
