+++ locales/hr.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Učitano iz lokalne pohrane",
   "{count} piece(s) of data detected": "{count} otkriveni dio/dijelovi podataka",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR kodovi su kreirani.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bitni znakovi — rade u običnom tekstu, e-pošti i terminalu.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, obrnuta polarnost (# je pozadina) — radi u običnom tekstu, e-pošti i terminalu.",
   "Add frame": "Dodaj okvir",
   "Address": "Adresa",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Zatvori dijalog",
   "Close guide": "Zatvori vodič",
   "Close scanner": "Zatvori skener",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktno: dva retka spakirana u jedan red pomoću znakova ▀ ▄ █.",
   "Configure frame settings": "Konfigurirajte postavke okvira",
   "Contact Card": "Kontaktna kartica",
-  "Copied": "Copied",
+  "Copied": "Kopirano",
   "Copy QR Code to clipboard": "Kopirajte QR kod u međuspremnik",
   "Copy to clipboard": "Kopiraj u međuspremnik",
   "Corners Dot color": "Kutovi Dot boja",
@@ -113,10 +113,10 @@
   "dots": "točkice",
   "Dots color": "Boja točkica",
   "Dots type": "Tip točkica",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Preuzmi .md",
+  "Download .txt": "Preuzmi .txt",
+  "Download all as .md": "Preuzmi sve kao .md",
+  "Download all as .txt": "Preuzmi sve kao .txt",
   "Download QR Code as JPG": "Preuzmi QR kod kao JPG",
   "Download QR Code as PNG": "Preuzmi QR kod kao PNG",
   "Download QR Code as SVG": "Preuzmi QR kod kao SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Primjer datoteke (vCard)",
   "Example with Custom Filenames": "Primjer s prilagođenim nazivima datoteka",
   "Export": "Izvoz",
-  "Export as text": "Export as text",
+  "Export as text": "Izvoz kao tekst",
   "Export QR code": "Izvezi QR kod",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Izvoz QR koda kao ASCII ili Unicode teksta",
   "extra-rounded": "izrazito zaobljen",
   "Failed to copy to clipboard": "Nije uspjelo kopirati u međuspremnik",
   "Failed to load changelog": "Neuspjelo učitavanje dnevnika promjena",
@@ -161,12 +161,12 @@
   "First Name": "Prvo ime",
   "First row preview:": "Pregled prve retka:",
   "Font family": "Obitelj fontova",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Oznaka okvira je isključena iz izvoza teksta — ASCII izlaz sadrži samo QR module.",
   "Frame preset": "Okvirna postavka",
   "Frame settings": "Postavke okvira",
   "Frame style": "Stil okvira",
   "Frame text": "Tekst okvira",
-  "Generating": "Generating",
+  "Generating": "Generiranje",
   "GitHub repository for this project": "GitHub repozitorij za ovaj projekt",
   "Have a beautiful day!": "Uživajte u prekrasnom danu!",
   "Have nice day!": "Ugodan dan!",
@@ -204,7 +204,7 @@
   "Message": "Poruka",
   "Mini QR Code Generator": "Mini generator QR kodova",
   "MiniQR's QR rendering engine has been updated": "Motor za renderiranje QR kodova MiniQR-a je ažuriran.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Većina QR skenera neće pročitati ovaj format jer znakovi # ostavljaju praznine između redaka. Koristite Unicode (polublokovi/cjeloblokovi) za skenirajući izvoz.",
   "Must be between 0 and 1": "Mora biti između 0 i 1",
   "Network Name (SSID)": "Naziv mreže (SSID)",
   "Network password (case-sensitive)": "Loyezak mreže (osjetljiv na velika i mala slova)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Molimo najprije otpremite datoteku.",
   "Position": "Pozicija",
   "Preset": "Prednastavka",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Pregled prve retka",
   "Processing...": "Obrada...",
   "QR Code Content": "Sadržaj QR koda",
   "QR code preset": "QR kod postavka",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Jednostavan URL/tekst",
   "Single export": "Izvoz pojedinačno",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Karakteristični puni blokovi (██) — najbolja vjernost u modernim fontovima.",
   "Sponsor": "Sponzor",
   "square": "kvadrat",
   "Start Date": "Datum početka",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Boja teksta",
   "Text is required": "Tekst je obavezan",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Tekstualni izlaz isključuje okvir i sliku.",
   "Text position": "Pozicija teksta",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Postoji previše QR kodova koji se istovremeno kopiraju u međuspremnik. Molimo preuzmite ih kao SVG ili PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Savjet: Za najbolje rezultate upotrijebite jasnu sliku s dobrim osvjetljenjem.",
   "Title": "Naslov",
   "Toggle dark mode": "Prekini tamni način rada",
   "top": "vrh",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unikod (puni blokovi)",
+  "Unicode (half-blocks)": "Unicode (polublokovi)",
   "Upload a CSV file": "Otpremi CSV datoteku",
   "Upload image": "Učitaj sliku",
   "Upload QR Code Image": "Otpremi sliku QR koda",
