+++ locales/hu.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Helyi tárolóból betöltve",
   "{count} piece(s) of data detected": "{count} az észlelt adat(ok)",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-kódokat hoztak létre.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7 bites karakterek – egyszerű szövegben, e-mailekben és terminálokon is működik.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bites, fordított polaritású (# jelenti a háttérszínt) — egyszerű szöveges üzenetekben, e-mailekben és terminálokon is működik.",
   "Add frame": "Keret hozzáadása",
   "Address": "Cím:",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Bezár párbeszédpanel",
   "Close guide": "Bezárás útmutató",
   "Close scanner": "Szkenner bezárása",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Tömör: két sor egyetlen sorba sűrítve a ▀ ▄ █ karakterek segítségével.",
   "Configure frame settings": "Keretbeállítások konfigurálása",
   "Contact Card": "Kapcsolat",
-  "Copied": "Copied",
+  "Copied": "Másolva",
   "Copy QR Code to clipboard": "QR-kód másolása a vágólapra",
   "Copy to clipboard": "Másolás a vágólapra",
   "Corners Dot color": "Sarkok Pöttyök színe",
@@ -113,10 +113,10 @@
   "dots": "pontok",
   "Dots color": "Pöttyök színe",
   "Dots type": "Pöttyök típusa",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".md fájl letöltése",
+  "Download .txt": ".txt fájl letöltése",
+  "Download all as .md": "Minden fájlt .md formátumban letölteni",
+  "Download all as .txt": "Minden fájl letöltése .txt formátumban",
   "Download QR Code as JPG": "QR-kód letöltése JPG-ként",
   "Download QR Code as PNG": "QR-kód letöltése PNG-ként",
   "Download QR Code as SVG": "QR-kód letöltése SVG-ként",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Példa fájl (vCard)",
   "Example with Custom Filenames": "Példa egyéni fájlnevekkel",
   "Export": "Export",
-  "Export as text": "Export as text",
+  "Export as text": "Szövegként exportálás",
   "Export QR code": "QR-kód exportálása",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "A QR-kód exportálása ASCII- vagy Unicode-szövegként",
   "extra-rounded": "extra kerek",
   "Failed to copy to clipboard": "Nem sikerült a vágólapra másolni",
   "Failed to load changelog": "Nem sikerült betölteni a changelogot",
@@ -161,12 +161,12 @@
   "First Name": "Keresztnév",
   "First row preview:": "Az első sor előnézete:",
   "Font family": "Betűcsalád",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "A keretcímke nem kerül bele a szöveges exportba – az ASCII-kimenet kizárólag a QR-modulokat tartalmazza.",
   "Frame preset": "Előre beállított keret",
   "Frame settings": "Keret beállítások",
   "Frame style": "Keret stílus",
   "Frame text": "Keret szöveg",
-  "Generating": "Generating",
+  "Generating": "Létrehozás",
   "GitHub repository for this project": "GitHub tároló ehhez a projekthez",
   "Have a beautiful day!": "Legyen szép napod!",
   "Have nice day!": "Szép napot!",
@@ -204,7 +204,7 @@
   "Message": "Üzenet",
   "Mini QR Code Generator": "Mini QR kód generátor",
   "MiniQR's QR rendering engine has been updated": "A MiniQR QR-megjelenítő motorja frissült",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "A legtöbb QR-olvasó nem tudja beolvasni ezt a formátumot, mivel a # karakterek szóközöket hagynak a sorok között. A beolvasható exportáláshoz használjon Unicode-ot (fél- vagy teljes blokkokat).",
   "Must be between 0 and 1": "0 és 1 közötti értéknek kell lennie",
   "Network Name (SSID)": "Hálózati név (SSID)",
   "Network password (case-sensitive)": "Hálózati jelszó (nagy- és kisbetű-érzékeny)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Kérjük, először töltsön fel egy fájlt",
   "Position": "Pozíció",
   "Preset": "Előre beállított",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Az első sor előnézete",
   "Processing...": "Feldolgozás...",
   "QR Code Content": "QR-kód tartalom",
   "QR code preset": "QR-kód előre beállított",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Egyszerű URL/szöveg",
   "Single export": "Egyszeri export",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Töltött négyzetek (██) — a legjobb megjelenés a modern betűtípusokban.",
   "Sponsor": "Szponzor",
   "square": "négyzet",
   "Start Date": "Kezdeti dátum",
@@ -291,15 +291,15 @@
   "Text": "Szöveg",
   "Text color": "Szöveg színe",
   "Text is required": "Szöveg szükséges",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "A szöveges kimenet nem tartalmazza a keretet és a képet",
   "Text position": "Szöveg pozíciója",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Túl sok QR-kódot lehet egyszerre a vágólapra másolni. Kérjük, töltse le őket inkább SVG vagy PNG formátumban.",
   "Tip: For best results, use a clear image with good lighting.": "Tipp: A legjobb eredmény elérése érdekében használjon tiszta képet jó megvilágítással.",
   "Title": "Cím",
   "Toggle dark mode": "Sötét mód bekapcsolása",
   "top": "teteje",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (teljes blokkok)",
+  "Unicode (half-blocks)": "Unicode (félblokkok)",
   "Upload a CSV file": "CSV fájl feltöltése",
   "Upload image": "Kép feltöltése",
   "Upload QR Code Image": "QR-kód kép feltöltése",
