+++ locales/id.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Dimuat dari penyimpanan lokal",
   "{count} piece(s) of data detected": "{count} potongan data yang terdeteksi",
   "{index} / {count} QR codes have been created.": "{index} / {count} Kode QR telah dibuat.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Karakter 7-bit — dapat digunakan dalam teks biasa, email, dan terminal.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, polaritas terbalik (# adalah latar belakang) — berfungsi di teks biasa, email, dan terminal.",
   "Add frame": "Tambahkan bingkai",
   "Address": "Alamat",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Menutup dialog",
   "Close guide": "Tutup panduan",
   "Close scanner": "Tutup pemindai",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Ringkas: dua baris digabungkan menjadi satu baris menggunakan karakter ▀ ▄ █.",
   "Configure frame settings": "Mengonfigurasi pengaturan bingkai",
   "Contact Card": "Kartu Kontak",
-  "Copied": "Copied",
+  "Copied": "Disalin",
   "Copy QR Code to clipboard": "Salin Kode QR ke papan klip",
   "Copy to clipboard": "Salin ke papan klip",
   "Corners Dot color": "Sudut Warna titik",
@@ -113,10 +113,10 @@
   "dots": "titik-titik",
   "Dots color": "Warna titik",
   "Dots type": "Jenis titik",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Unduh .md",
+  "Download .txt": "Unduh berkas .txt",
+  "Download all as .md": "Unduh semuanya dalam format .md",
+  "Download all as .txt": "Unduh semuanya dalam format .txt",
   "Download QR Code as JPG": "Unduh Kode QR sebagai JPG",
   "Download QR Code as PNG": "Unduh Kode QR sebagai PNG",
   "Download QR Code as SVG": "Unduh Kode QR sebagai SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Contoh file (vCard)",
   "Example with Custom Filenames": "Contoh dengan Nama File Kustom",
   "Export": "Ekspor",
-  "Export as text": "Export as text",
+  "Export as text": "Ekspor sebagai teks",
   "Export QR code": "Ekspor kode QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Ekspor Kode QR sebagai teks ASCII atau Unicode",
   "extra-rounded": "ekstra bulat",
   "Failed to copy to clipboard": "Gagal menyalin ke papan klip",
   "Failed to load changelog": "Gagal memuat log perubahan",
@@ -161,12 +161,12 @@
   "First Name": "Nama depan",
   "First row preview:": "Pratinjau baris pertama:",
   "Font family": "Keluarga huruf",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Label bingkai tidak disertakan dalam ekspor teks — hasil keluaran ASCII hanya berisi modul QR.",
   "Frame preset": "Preset bingkai",
   "Frame settings": "Pengaturan bingkai",
   "Frame style": "Gaya bingkai",
   "Frame text": "Bingkai teks",
-  "Generating": "Generating",
+  "Generating": "Membuat",
   "GitHub repository for this project": "Repositori GitHub untuk proyek ini",
   "Have a beautiful day!": "Selamat menikmati hari yang indah!",
   "Have nice day!": "Semoga harimu menyenangkan!",
@@ -204,7 +204,7 @@
   "Message": "Pesan",
   "Mini QR Code Generator": "Generator Kode QR Mini",
   "MiniQR's QR rendering engine has been updated": "Mesin rendering QR MiniQR telah diperbarui",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Sebagian besar pemindai QR tidak dapat membaca format ini karena simbol # meninggalkan celah di antara baris. Gunakan Unicode (blok setengah/penuh) agar hasil ekspornya dapat dipindai.",
   "Must be between 0 and 1": "Harus berada di antara 0 dan 1",
   "Network Name (SSID)": "Nama Jaringan (SSID)",
   "Network password (case-sensitive)": "Kata sandi jaringan (peka huruf besar/kecil)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Silakan unggah file terlebih dahulu",
   "Position": "Posisi",
   "Preset": "Pengaturan default",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Tampilan pratinjau baris pertama",
   "Processing...": "Pemrosesan...",
   "QR Code Content": "Konten Kode QR",
   "QR code preset": "Preset kode QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL Sederhana/Teks",
   "Single export": "Ekspor tunggal",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Karakter blok solid (██) — kualitas terbaik pada font modern.",
   "Sponsor": "Sponsor",
   "square": "persegi",
   "Start Date": "Tanggal Mulai",
@@ -291,15 +291,15 @@
   "Text": "Teks",
   "Text color": "Warna teks",
   "Text is required": "Diperlukan teks",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Teks yang ditampilkan tidak mencakup bingkai dan gambar",
   "Text position": "Posisi teks",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Ada terlalu banyak kode QR yang harus disalin ke clipboard sekaligus. Silakan unduh sebagai SVG atau PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Saran: Untuk hasil terbaik, gunakan gambar yang jernih dengan pencahayaan yang bagus.",
   "Title": "Judul",
   "Toggle dark mode": "Aktifkan mode gelap",
   "top": "atas",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blok lengkap)",
+  "Unicode (half-blocks)": "Unicode (blok setengah)",
   "Upload a CSV file": "Unggah file CSV",
   "Upload image": "Unggah gambar",
   "Upload QR Code Image": "Unggah Gambar Kode QR",
