+++ locales/it.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Ultimo salvataggio locale",
   "{count} piece(s) of data detected": "{count} pezzo(i) di dati rilevati",
   "{index} / {count} QR codes have been created.": "{index} / {count} codici QR sono stati creati.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Caratteri a 7 bit: funzionano nel testo semplice, nelle e-mail e nei terminali.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bit, polarità invertita (# indica lo sfondo) — funziona in testo semplice, nelle e-mail e nei terminali.",
   "Add frame": "Aggiungi cornice",
   "Address": "Indirizzo",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Chiudere la finestra di dialogo",
   "Close guide": "Chiudi guida",
   "Close scanner": "Chiudi scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compatto: due righe raggruppate in una sola riga utilizzando i simboli ▀ ▄ █.",
   "Configure frame settings": "Configurare le impostazioni della cornice",
   "Contact Card": "Scheda Contatto",
-  "Copied": "Copied",
+  "Copied": "Copiato",
   "Copy QR Code to clipboard": "Copia il codice QR negli appunti",
   "Copy to clipboard": "Copia negli appunti",
   "Corners Dot color": "Colore dei punti agli angoli",
@@ -113,10 +113,10 @@
   "dots": "punti",
   "Dots color": "Colore dei punti",
   "Dots type": "Stile dei punti",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Scarica .md",
+  "Download .txt": "Scarica il file .txt",
+  "Download all as .md": "Scarica tutto in formato .md",
+  "Download all as .txt": "Scarica tutto in formato .txt",
   "Download QR Code as JPG": "Scarica il codice QR come JPG",
   "Download QR Code as PNG": "Scarica il codice QR come PNG",
   "Download QR Code as SVG": "Scarica il codice QR come SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "File di esempio (vCard)",
   "Example with Custom Filenames": "Esempio con nomi file personalizzati",
   "Export": "Esporta",
-  "Export as text": "Export as text",
+  "Export as text": "Esporta come testo",
   "Export QR code": "Esporta codice QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Esporta il codice QR come testo ASCII o Unicode",
   "extra-rounded": "extra arrotondato",
   "Failed to copy to clipboard": "Impossibile copiare negli appunti",
   "Failed to load changelog": "Caricamento del changelog fallito",
@@ -161,12 +161,12 @@
   "First Name": "Nome",
   "First row preview:": "Anteprima della prima riga:",
   "Font family": "Famiglia di caratteri",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "L'etichetta del fotogramma è esclusa dall'esportazione del testo: l'output ASCII contiene solo i moduli QR.",
   "Frame preset": "Cornice preimpostata",
   "Frame settings": "Impostazioni cornice",
   "Frame style": "Stile cornice",
   "Frame text": "Testo della cornice",
-  "Generating": "Generating",
+  "Generating": "Generazione",
   "GitHub repository for this project": "Repository GitHub per questo progetto",
   "Have a beautiful day!": "Buona giornata!",
   "Have nice day!": "Buona giornata!",
@@ -204,7 +204,7 @@
   "Message": "Messaggio",
   "Mini QR Code Generator": "Mini generatore di codici QR",
   "MiniQR's QR rendering engine has been updated": "Il motore di rendering QR di MiniQR è stato aggiornato",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "La maggior parte degli scanner di codici QR non legge questo formato perché i simboli # lasciano degli spazi tra le righe. Per un'esportazione leggibile dallo scanner, usa l'Unicode (blocchi parziali o completi).",
   "Must be between 0 and 1": "Deve essere compreso tra 0 e 1",
   "Network Name (SSID)": "Nome della rete (SSID)",
   "Network password (case-sensitive)": "Password di rete (sensibile alle maiuscole)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Carica prima un file",
   "Position": "Posizione",
   "Preset": "Modello",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Anteprima della prima riga",
   "Processing...": "Elaborando…",
   "QR Code Content": "Contenuto del codice QR",
   "QR code preset": "Modello del codice QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/Testo semplice",
   "Single export": "Esportazione singola",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Caratteri a blocchi pieni (██) — la massima fedeltà nei font moderni.",
   "Sponsor": "Sponsor",
   "square": "quadrato",
   "Start Date": "Data di inizio",
@@ -291,15 +291,15 @@
   "Text": "Testo",
   "Text color": "Colore testo",
   "Text is required": "Il testo è obbligatorio",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Il testo visualizzato non include la cornice e l'immagine",
   "Text position": "Posizione testo",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Ci sono troppi codici QR per copiarli tutti negli appunti. Si prega di scaricarli come SVG o PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Suggerimento: per ottenere risultati ottimali, utilizzare un'immagine chiara e ben illuminata.",
   "Title": "Titolo",
   "Toggle dark mode": "Attiva/disattiva modalità scura",
   "top": "alto",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blocchi completi)",
+  "Unicode (half-blocks)": "Unicode (mezzi blocchi)",
   "Upload a CSV file": "Caricare un file CSV",
   "Upload image": "Carica immagine",
   "Upload QR Code Image": "Carica immagine di un codice QR",
