+++ locales/ja.json
@@ -51,8 +51,8 @@
   "Last saved locally": "ローカルストレージから読み込みました",
   "{count} piece(s) of data detected": "{count} 個のデータが検出されました",
   "{index} / {count} QR codes have been created.": "{index} / {count} QRコードが作成されました。",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7ビット文字 — プレーンテキスト、電子メール、およびターミナルで動作します。",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7ビット、極性反転（#が背景色） — プレーンテキスト、電子メール、およびターミナルで動作します。",
   "Add frame": "枠を追加",
   "Address": "住所",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "ダイアログを閉じる",
   "Close guide": "ガイドを閉じる",
   "Close scanner": "スキャナーを閉じる",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "コンパクト：▀ ▄ █ の文字を使って、2列を1行にまとめました。",
   "Configure frame settings": "フレームの設定",
   "Contact Card": "連絡先カード",
-  "Copied": "Copied",
+  "Copied": "コピーしました",
   "Copy QR Code to clipboard": "QRコードをクリップボードにコピー",
   "Copy to clipboard": "クリップボードにコピー",
   "Corners Dot color": "角のドットの色",
@@ -113,10 +113,10 @@
   "dots": "ドット",
   "Dots color": "ドットの色",
   "Dots type": "ドットの種類",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".mdファイルをダウンロード",
+  "Download .txt": ".txtファイルをダウンロード",
+  "Download all as .md": "すべてを.md形式でダウンロード",
+  "Download all as .txt": "すべてを.txt形式でダウンロード",
   "Download QR Code as JPG": "QRコードをJPGでダウンロード",
   "Download QR Code as PNG": "QRコードをPNG形式でダウンロード",
   "Download QR Code as SVG": "QRコードをSVGとしてダウンロード",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "ファイル例（vCard）",
   "Example with Custom Filenames": "カスタムファイル名を使用した例",
   "Export": "エクスポート",
-  "Export as text": "Export as text",
+  "Export as text": "テキストとしてエクスポート",
   "Export QR code": "QRコードをエクスポート",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QRコードをASCIIまたはUnicodeテキストとしてエクスポートする",
   "extra-rounded": "余分な丸み",
   "Failed to copy to clipboard": "クリップボードにコピーできませんでした",
   "Failed to load changelog": "変更履歴の読み込みに失敗しました",
@@ -161,12 +161,12 @@
   "First Name": "名前",
   "First row preview:": "最初の行プレビュー:",
   "Font family": "フォントファミリー",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "フレームラベルはテキストのエクスポート対象外です。ASCII形式の出力にはQRモジュールのみが含まれます。",
   "Frame preset": "フレームプリセット",
   "Frame settings": "フレーム設定",
   "Frame style": "枠のスタイル",
   "Frame text": "フレームテキスト",
-  "Generating": "Generating",
+  "Generating": "生成",
   "GitHub repository for this project": "このプロジェクトの GitHub リポジトリ",
   "Have a beautiful day!": "良い一日を！",
   "Have nice day!": "楽しい一日をお過ごしください。",
@@ -204,7 +204,7 @@
   "Message": "メッセージ",
   "Mini QR Code Generator": "ミニQRコードジェネレーター",
   "MiniQR's QR rendering engine has been updated": "MiniQRのQRコード描画エンジンが更新されました",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "「#」記号によって行間にスペースが入るため、ほとんどのQRコードリーダーはこの形式を読み取ることができません。スキャン可能な形式でエクスポートするには、Unicode（ハーフブロックまたはフルブロック）を使用してください。",
   "Must be between 0 and 1": "0から1の間でなければなりません",
   "Network Name (SSID)": "ネットワーク名（SSID）",
   "Network password (case-sensitive)": "ネットワークパスワード（大文字と小文字を区別）",
@@ -243,7 +243,7 @@
   "Please upload a file first": "最初にファイルをアップロードしてください",
   "Position": "位置",
   "Preset": "プリセット",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "1行目のプレビュー",
   "Processing...": "処理中...",
   "QR Code Content": "QRコードの内容",
   "QR code preset": "QRコードのプリセット",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "シンプルなURL/テキスト",
   "Single export": "単一エクスポート",
   "SMS": "ショートメール",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "実線ブロック文字（██） — 現代のフォントにおいて最も忠実な再現が可能です。",
   "Sponsor": "スポンサー情報",
   "square": "正方形の",
   "Start Date": "開始日",
@@ -291,15 +291,15 @@
   "Text": "テキスト",
   "Text color": "テキストの色",
   "Text is required": "テキストが必要",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "テキスト出力には、フレームと画像は含まれません",
   "Text position": "テキストの位置",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "一度にクリップボードにコピーできるQRコードが多すぎます。SVGまたはPNGとしてダウンロードしてください。",
   "Tip: For best results, use a clear image with good lighting.": "ヒント: 最良の結果を得るには、良い照明で明確な画像を使用してください。",
   "Title": "タイトル",
   "Toggle dark mode": "ダークモードの切り替え",
   "top": "トップ",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode（フルブロック）",
+  "Unicode (half-blocks)": "Unicode（ハーフブロック）",
   "Upload a CSV file": "CSVファイルのアップロード",
   "Upload image": "画像をアップロード",
   "Upload QR Code Image": "QRコード画像をアップロード",
