+++ locales/ko.json
@@ -51,8 +51,8 @@
   "Last saved locally": "로컬 스토리지에서 로드",
   "{count} piece(s) of data detected": "{count} 감지된 데이터 조각",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR 코드가 생성되었습니다.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7비트 문자 — 일반 텍스트, 이메일 및 터미널에서 작동합니다.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7비트, 반전된 극성(#은 배경색) — 일반 텍스트, 이메일 및 터미널에서 작동합니다.",
   "Add frame": "프레임 추가",
   "Address": "주소",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "대화 상자 닫기",
   "Close guide": "가이드 닫기",
   "Close scanner": "스캐너 닫기",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "압축: ▀, ▄, █ 문자를 사용하여 두 행을 한 줄로 압축했습니다.",
   "Configure frame settings": "프레임 설정 구성",
   "Contact Card": "연락처 카드",
-  "Copied": "Copied",
+  "Copied": "복사됨",
   "Copy QR Code to clipboard": "QR 코드를 클립보드에 복사",
   "Copy to clipboard": "클립보드에 복사",
   "Corners Dot color": "모서리 도트 색상",
@@ -113,10 +113,10 @@
   "dots": "점",
   "Dots color": "점 색상",
   "Dots type": "점 유형",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".md 파일 다운로드",
+  "Download .txt": ".txt 파일 다운로드",
+  "Download all as .md": "모두 .md 파일로 다운로드",
+  "Download all as .txt": "모두 .txt 파일로 다운로드",
   "Download QR Code as JPG": "QR 코드를 JPG로 다운로드",
   "Download QR Code as PNG": "QR 코드를 PNG로 다운로드",
   "Download QR Code as SVG": "QR 코드를 SVG로 다운로드",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "예제 파일(vCard)",
   "Example with Custom Filenames": "맞춤 파일 이름 예시",
   "Export": "수출",
-  "Export as text": "Export as text",
+  "Export as text": "텍스트로 내보내기",
   "Export QR code": "QR 코드 내보내기",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR 코드를 ASCII 또는 유니코드 텍스트로 내보내기",
   "extra-rounded": "추가 라운드",
   "Failed to copy to clipboard": "클립보드에 복사하지 못했습니다.",
   "Failed to load changelog": "변경 로그를 로드하지 못했습니다.",
@@ -161,12 +161,12 @@
   "First Name": "이름",
   "First row preview:": "첫 번째 행 미리 보기:",
   "Font family": "글꼴 가족",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "프레임 라벨은 텍스트 내보내기에서 제외됩니다. ASCII 출력에는 QR 모듈만 포함됩니다.",
   "Frame preset": "프레임 사전 설정",
   "Frame settings": "프레임 설정",
   "Frame style": "프레임 스타일",
   "Frame text": "프레임 텍스트",
-  "Generating": "Generating",
+  "Generating": "생성",
   "GitHub repository for this project": "이 프로젝트의 GitHub 리포지토리",
   "Have a beautiful day!": "아름다운 하루 되세요!",
   "Have nice day!": "좋은 하루 되세요!",
@@ -204,7 +204,7 @@
   "Message": "메시지",
   "Mini QR Code Generator": "미니 QR코드 생성기",
   "MiniQR's QR rendering engine has been updated": "MiniQR의 QR 렌더링 엔진이 업데이트되었습니다",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "대부분의 QR 스캐너는 # 기호가 행 사이에 공백을 남기기 때문에 이 형식을 인식하지 못합니다. 스캔 가능한 파일로 내보내려면 유니코드(하프/풀 블록)를 사용하세요.",
   "Must be between 0 and 1": "0과 1 사이여야 합니다",
   "Network Name (SSID)": "네트워크 이름(SSID)",
   "Network password (case-sensitive)": "네트워크 비밀번호(대소문자 구분)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "먼저 파일을 업로드하세요.",
   "Position": "직위",
   "Preset": "프리셋",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "첫 번째 행 미리 보기",
   "Processing...": "처리 중...",
   "QR Code Content": "QR코드 콘텐츠",
   "QR code preset": "QR코드 사전 설정",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "간단한 URL/텍스트",
   "Single export": "단일 내보내기",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "실선 블록 문자 (██) — 최신 폰트에서 가장 정확한 재현을 제공합니다.",
   "Sponsor": "후원사",
   "square": "정사각형",
   "Start Date": "시작 날짜",
@@ -291,15 +291,15 @@
   "Text": "텍스트",
   "Text color": "텍스트 색상",
   "Text is required": "텍스트 입력 필수",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "텍스트 출력에는 프레임과 이미지가 포함되지 않습니다",
   "Text position": "텍스트 위치",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "QR 코드가 너무 많아 클립보드에 한 번에 복사할 수 없습니다. 대신 SVG 또는 PNG로 다운로드하세요.",
   "Tip: For best results, use a clear image with good lighting.": "팁: 팁: 최상의 결과를 얻으려면 조명이 밝고 선명한 이미지를 사용하세요.",
   "Title": "제목",
   "Toggle dark mode": "다크 모드 전환",
   "top": "상단",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "유니코드 (전체 블록)",
+  "Unicode (half-blocks)": "유니코드 (반블록)",
   "Upload a CSV file": "CSV 파일 업로드",
   "Upload image": "이미지 업로드",
   "Upload QR Code Image": "QR코드 이미지 업로드",
