+++ locales/lt.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Įkelta iš vietinės saugyklos",
   "{count} piece(s) of data detected": "{count} aptikta (-os) duomenų dalis (-ys)",
   "{index} / {count} QR codes have been created.": "{index} / {count} Sukurti QR kodai.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7 bitų simboliai — tinka paprastam tekstui, elektroniniams laiškams ir terminalams.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bitų, apversta polaritetas (simbolis „#“ reiškia foną) — veikia paprastame tekste, el. laiškuose ir terminaluose.",
   "Add frame": "Pridėti rėmelį",
   "Address": "Adresas",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Uždaryti dialogo langą",
   "Close guide": "Uždaryti vadovą",
   "Close scanner": "Uždaryti skaitytuvą",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktiškas: dvi eilės, išdėstytos vienoje eilutėje naudojant simbolius ▀ ▄ █.",
   "Configure frame settings": "Sukonfigūruokite rėmelio nustatymus",
   "Contact Card": "Kontaktinė kortelė",
-  "Copied": "Copied",
+  "Copied": "Nukopijuota",
   "Copy QR Code to clipboard": "QR kodo kopijavimas į iškarpinę",
   "Copy to clipboard": "Kopijuoti į iškarpinę",
   "Corners Dot color": "Kampai Taškų spalva",
@@ -113,10 +113,10 @@
   "dots": "taškai",
   "Dots color": "Taškų spalva",
   "Dots type": "Taškų tipas",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Atsisiųsti .md",
+  "Download .txt": "Atsisiųsti .txt",
+  "Download all as .md": "Atsisiųsti viską kaip .md",
+  "Download all as .txt": "Atsisiųsti viską kaip .txt failą",
   "Download QR Code as JPG": "Atsisiųsti QR kodą kaip JPG",
   "Download QR Code as PNG": "Atsisiųsti QR kodą kaip PNG",
   "Download QR Code as SVG": "Atsisiųsti QR kodą kaip SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Failo pavyzdys (vCard)",
   "Example with Custom Filenames": "Pavyzdys su pasirinktiniais failų pavadinimais",
   "Export": "Eksportuoti",
-  "Export as text": "Export as text",
+  "Export as text": "Eksportuoti kaip tekstą",
   "Export QR code": "Eksportuoti QR kodą",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksportuoti QR kodą kaip ASCII arba Unicode tekstą",
   "extra-rounded": "itin užapvalintas",
   "Failed to copy to clipboard": "Nepavyko nukopijuoti į iškarpinę",
   "Failed to load changelog": "Nepavyko įkelti pakeitimų žurnalo",
@@ -161,12 +161,12 @@
   "First Name": "Vardas",
   "First row preview:": "Pirmosios eilės peržiūra:",
   "Font family": "Šriftų šeima",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Rėmelio etiketė neįtraukiama į eksportuojamą tekstą – ASCII formatu pateikiami tik QR moduliai.",
   "Frame preset": "Išankstinis kadro nustatymas",
   "Frame settings": "Rėmo nustatymai",
   "Frame style": "Rėmo stilius",
   "Frame text": "Rėmelio tekstas",
-  "Generating": "Generating",
+  "Generating": "Kūrimas",
   "GitHub repository for this project": "Šio projekto \"GitHub\" saugykla",
   "Have a beautiful day!": "Gražios dienos!",
   "Have nice day!": "Geros dienos!",
@@ -204,7 +204,7 @@
   "Message": "Žinutė",
   "Mini QR Code Generator": "Mini QR kodų generatorius",
   "MiniQR's QR rendering engine has been updated": "Atnaujintas „MiniQR“ QR kodų atvaizdavimo variklis",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Dauguma QR kodų skaitytuvų šio formato neskaito, nes simboliai „#“ palieka tarpelius tarp eilučių. Norėdami eksportuoti failą, kurį būtų galima nuskaityti, naudokite Unicode (pusinius arba pilnus blokus).",
   "Must be between 0 and 1": "Turi būti nuo 0 iki 1",
   "Network Name (SSID)": "Tinklo pavadinimas (SSID)",
   "Network password (case-sensitive)": "Tinklo slaptažodis (priklauso nuo mažųjų raidžių)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Pirmiausia įkelkite failą",
   "Position": "Pozicija",
   "Preset": "Iš anksto nustatyta",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Pirmojo eilutės peržiūra",
   "Processing...": "Apdorojimas...",
   "QR Code Content": "QR kodo turinys",
   "QR code preset": "QR kodo išankstinis nustatymas",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Paprastas URL/tekstas",
   "Single export": "Vienkartinis eksportas",
   "SMS": "trumpasis pranešimas",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Vienalytės blokinės raidės (██) — geriausias atvaizdavimas šiuolaikiniuose šriftuose.",
   "Sponsor": "Rėmėjas",
   "square": "kvadratas",
   "Start Date": "Pradžios data",
@@ -291,15 +291,15 @@
   "Text": "Tekstas",
   "Text color": "Teksto spalva",
   "Text is required": "Būtina pateikti tekstą",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Teksto išvestyje neatsižvelgiama į rėmelį ir paveikslėlį",
   "Text position": "Teksto padėtis",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Vienu metu į iškarpinę galima nukopijuoti per daug QR kodų. Vietoj to atsisiųskite juos kaip SVG arba PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Patarimas: Norėdami gauti geriausius rezultatus, naudokite aiškų vaizdą su geru apšvietimu.",
   "Title": "Pavadinimas",
   "Toggle dark mode": "Perjungti tamsų režimą",
   "top": "viršuje",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (pilni blokai)",
+  "Unicode (half-blocks)": "Unicode (pusblokiai)",
   "Upload a CSV file": "Įkelkite CSV failą",
   "Upload image": "Įkelti paveikslėlį",
   "Upload QR Code Image": "Įkelti QR kodo paveikslėlį",
