+++ locales/ms.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Dimuat daripada storan tempatan",
   "{count} piece(s) of data detected": "{count} Potongan data dikesan",
   "{index} / {count} QR codes have been created.": "{index} / {count} Kod QR telah dihasilkan.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Karakter 7-bit — berfungsi dalam teks biasa, e-mel, dan terminal.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, polariti terbalik (# adalah latar belakang) — berfungsi dalam teks biasa, e-mel, dan terminal.",
   "Add frame": "Tambah bingkai",
   "Address": "Alamat",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Tutup dialog",
   "Close guide": "Tutup panduan",
   "Close scanner": "Tutup imbas",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Padat: dua baris disusun dalam satu baris menggunakan aksara ▀ ▄ █.",
   "Configure frame settings": "Konfigurasikan tetapan bingkai",
   "Contact Card": "Kad Kenalan",
-  "Copied": "Copied",
+  "Copied": "Disalin",
   "Copy QR Code to clipboard": "Salin Kod QR ke papan klip",
   "Copy to clipboard": "Salin ke papan pemotong",
   "Corners Dot color": "Titiskan warna pada sudut",
@@ -113,10 +113,10 @@
   "dots": "titik-titik",
   "Dots color": "Warna titik",
   "Dots type": "Jenis titik",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Muat turun .md",
+  "Download .txt": "Muat turun .txt",
+  "Download all as .md": "Muat turun semuanya sebagai .md",
+  "Download all as .txt": "Muat turun semuanya sebagai .txt",
   "Download QR Code as JPG": "Muat turun Kod QR sebagai JPG",
   "Download QR Code as PNG": "Muat turun Kod QR sebagai PNG",
   "Download QR Code as SVG": "Muat turun Kod QR sebagai SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Fail contoh (vCard)",
   "Example with Custom Filenames": "Contoh dengan Nama Fail Tersuai",
   "Export": "Eksport",
-  "Export as text": "Export as text",
+  "Export as text": "Eksport sebagai teks",
   "Export QR code": "Eksport kod QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksport Kod QR sebagai teks ASCII atau Unicode",
   "extra-rounded": "sangat bulat",
   "Failed to copy to clipboard": "Gagal menyalin ke papan keratan",
   "Failed to load changelog": "Gagal memuatkan log perubahan",
@@ -161,12 +161,12 @@
   "First Name": "Nama Pertama",
   "First row preview:": "Preview barisan pertama:",
   "Font family": "Keluarga fon",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Label bingkai dikecualikan daripada eksport teks — keluaran ASCII hanya mengandungi modul QR.",
   "Frame preset": "Praaturan bingkai",
   "Frame settings": "Tetapan bingkai",
   "Frame style": "Gaya bingkai",
   "Frame text": "Teks bingkai",
-  "Generating": "Generating",
+  "Generating": "Menghasilkan",
   "GitHub repository for this project": "Repositori GitHub untuk projek ini",
   "Have a beautiful day!": "Semoga hari anda indah!",
   "Have nice day!": "Semoga hari anda menyenangkan!",
@@ -204,7 +204,7 @@
   "Message": "Mesej",
   "Mini QR Code Generator": "Penjana Kod QR Mini",
   "MiniQR's QR rendering engine has been updated": "Enjin pemaparan QR MiniQR telah dikemas kini.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Kebanyakan pemindai QR tidak akan membaca format ini kerana glif # meninggalkan jurang antara baris. Gunakan Unicode (blok separuh/penuh) untuk eksport yang boleh diimbas.",
   "Must be between 0 and 1": "Mesti antara 0 dan 1",
   "Network Name (SSID)": "Nama Rangkaian (SSID)",
   "Network password (case-sensitive)": "Kata laluan rangkaian (peka huruf besar/kecil)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Sila muat naik fail terlebih dahulu.",
   "Position": "Jawatan",
   "Preset": "Pra-tetapan",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Pratinjau baris pertama",
   "Processing...": "Pemprosesan...",
   "QR Code Content": "Kandungan Kod QR",
   "QR code preset": "Prset kod QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/Teks Mudah",
   "Single export": "Eksport tunggal",
   "SMS": "Mesej ringkas",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Karakter blok pepejal (██) — kesetiaan terbaik dalam fon moden.",
   "Sponsor": "Penaja",
   "square": "petak",
   "Start Date": "Tarikh Mula",
@@ -291,15 +291,15 @@
   "Text": "Teks",
   "Text color": "Warna teks",
   "Text is required": "Teks diperlukan",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Keluaran teks tidak termasuk bingkai dan imej.",
   "Text position": "Posisi teks",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Terdapat terlalu banyak kod QR yang hendak disalin ke papan klip sekaligus. Sila muat turun mereka sebagai SVG atau PNG sebaliknya.",
   "Tip: For best results, use a clear image with good lighting.": "Petua: Untuk hasil terbaik, gunakan imej yang jelas dengan pencahayaan yang baik.",
   "Title": "Tajuk",
   "Toggle dark mode": "Togol mod gelap",
   "top": "atas",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unikod (blok penuh)",
+  "Unicode (half-blocks)": "Unikod (separuh blok)",
   "Upload a CSV file": "Muat naik fail CSV",
   "Upload image": "Muat naik imej",
   "Upload QR Code Image": "Muat naik imej kod QR",
