+++ locales/my.json
@@ -51,11 +51,11 @@
   "Last saved locally": "လိုကယ် သိုလှောင်မှုမှ ဝန်ဆွဲတင်ခဲ့သည်",
   "{count} piece(s) of data detected": "ဒေတာ အပိုင်းအစ {count} ခုအား တွေ့ရှိသည်",
   "{index} / {count} QR codes have been created.": "{index} / {count} ကျူအာကုဒ်များကို ဖန်တီးပြီးပါပြီ။",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "၇-ဘစ် စာလုံးများ — ပလိန်း တက်စ်၊ အီးမေးလ်နှင့် တိနယ်လ်များတွင် အသုံးပြုနိုင်သည်။",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-ဘစ်၊ အနုတ်ဘက်ပြောင်းထားသော ပိုလာရီတီ (# သည် နောက်ခံဖြစ်သည်) — ပလိန်းတက်စ်၊ အီးမေးလ်နှင့် တာမီနယ်များတွင် အလုပ်လုပ်သည်။",
   "Add frame": "ဖရိန် ထည့်ရန်",
   "Address": "လိပ်စာ",
-  "ASCII": "ASCII",
+  "ASCII": "အက်စကီး",
   "Available Fields": "ရနိုင်သော နေရာများ",
   "Background color": "နောက်ခံအရောင်",
   "Batch export": "အစုလိုက် တင်ပို့ရန်",
@@ -82,10 +82,10 @@
   "Close dialog": " ဒိုင်ယာလော့ ပိတ်ရန်",
   "Close guide": "လမ်းညွှန် ပိတ်ရန်",
   "Close scanner": "စကင်နာကို ပိတ်ရန်",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: ▀ ▄ █ သင်္ကေတများဖြင့် အတန်းနှစ်ခုကို တစ်ကြောင်းထဲသို့ ထည့်ထားသည်။",
   "Configure frame settings": "ဖရိန်အပြင်အဆင် သတ်မှတ်ရန်",
   "Contact Card": "လိပ်စာကဒ်",
-  "Copied": "Copied",
+  "Copied": "ကူးယူထားသည်",
   "Copy QR Code to clipboard": "ကလစ်ဘုတ်သို့ ကျူအာကုဒ်အား ကူးယူရန်",
   "Copy to clipboard": "ကလစ်ဘုတ်သို့ ကူးယူရန်",
   "Corners Dot color": "ထောင့်အစက်အပြောက် အရောင်",
@@ -113,10 +113,10 @@
   "dots": "အစက်အပြောက်များ",
   "Dots color": "အစက်အပြောက်များ အရောင်",
   "Dots type": "အစက်အပြောက် အမျိုးအစား",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "ဒေါင်းလုဒ် .md",
+  "Download .txt": ".txt ကို ဒေါင်းလုဒ်လုပ်ပါ။",
+  "Download all as .md": "အားလုံးကို .md ဖိုင်အဖြစ် ဒေါင်းလုဒ်ဆွဲပါ။",
+  "Download all as .txt": "အားလုံးကို .txt ဖိုင်အဖြစ် ဒေါင်းလုဒ်ဆွဲပါ။",
   "Download QR Code as JPG": "JPG အနေဖြင့် ကျူအာကုဒ်အား ဒေါင်းလုပ်ရန်",
   "Download QR Code as PNG": "PNG အနေဖြင့် ကျူအာကုဒ်အား ဒေါင်းလုပ်ရန်",
   "Download QR Code as SVG": "SVG အနေဖြင့် ကျူအာကုဒ်အား ဒေါင်းလုပ်ရန်",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "နမူနာဖိုင် (vCard)",
   "Example with Custom Filenames": "စိတ်ကြိုက် ဖိုင်အမည်များပါသည့် ဥပမာ",
   "Export": "တင်ပို့ရန်",
-  "Export as text": "Export as text",
+  "Export as text": "စာသားအဖြစ် ထုတ်ပို့ပါ",
   "Export QR code": "ကျူအာကုဒ်အား တင်ပို့ရန်",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR ကုဒ်ကို ASCII သို့မဟုတ် Unicode စာသားအဖြစ် ထုတ်ပို့ပါ။",
   "extra-rounded": "အနားသတ် ပိုမိုလုံး",
   "Failed to copy to clipboard": "ကလစ်ဘုတ်သို့ ကူးယူခြင်း မအောင်မြင်ပါ",
   "Failed to load changelog": "ပြောင်းလဲမှု မှတ်တမ်း ရယူရန် မအောင်မြင်ပါ",
@@ -161,12 +161,12 @@
   "First Name": "ပထမအမည်",
   "First row preview:": "ပထမအတန်း နမူနာ:",
   "Font family": "စာလုံးပုံစံ မိသားစု",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "ဖရိမ်အမှတ်အသားကို စာသားထုတ်ယူမှုမှ ထုတ်ပယ်ထားသည် — ASCII ထုတ်လွှတ်ချက်တွင် QR မော်ဂျူးများသာ ပါဝင်သည်။",
   "Frame preset": "ဖရိန် ကြိုတင်သတ်မှတ်မှု",
   "Frame settings": "ဘောင် အပြင်အဆင်များ",
   "Frame style": "ဖရိန် ပုံစံ",
   "Frame text": "ဖရိန် စာသား",
-  "Generating": "Generating",
+  "Generating": "ထုတ်လုပ်ခြင်း",
   "GitHub repository for this project": "ဤပရောဂျက်၏ ဂစ်ဟက် သိုလှောင်ရုံ",
   "Have a beautiful day!": "ကောင်းသောနေ့ ဖြစ်ပါစေ။",
   "Have nice day!": "ကောင်းသောနေ့ပါ။",
@@ -204,7 +204,7 @@
   "Message": "စာလွှာ",
   "Mini QR Code Generator": "မီနီကျူအာ ထုတ်လုပ်သူ",
   "MiniQR's QR rendering engine has been updated": "MiniQR ၏ QR ဖော်ပြမှု အင်ဂျင်ကို အပ်ဒိတ်လုပ်ထားသည်။",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "အများစု QR စကင်နာများသည် # သင်္ကေတများကြောင့် အတန်းများကြားတွင် အကွာအဝေးများ ကျန်ရှိနေသောကြောင့် ဤဖော်မတ်ကို ဖတ်မရနိုင်ပါ။ စကင်နိုင်သော ထုတ်ပို့မှုအတွက် Unicode (ဝက်ဘလော့ခ်/ပြည့်စုံဘလော့ခ်များ) ကို အသုံးပြုပါ။",
   "Must be between 0 and 1": "0 နှင့် 1 ကြား ဖြစ်ရမည်။",
   "Network Name (SSID)": "ကွန်ရက်အမည် (SSID)",
   "Network password (case-sensitive)": "ကွန်ရက် စကားဝှက် (စာလုံးအကြီးအသေးအတိုင်း)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "ကျေးဇူးပြု၍ ဖိုင်တစ်ဖိုင်အား တင်ပါ",
   "Position": "ရာထူး",
   "Preset": "ပုံစံ",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "ပထမတန်းကြိုကြည့်ခြင်း",
   "Processing...": "လုပ်ဆောင်နေစဉ် ...",
   "QR Code Content": "ကျူအာကုဒ် အကြောင်းအရာ",
   "QR code preset": "ကျူအာကုဒ် ပုံစံများ",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "ရိုးရှင်း URL/စာသား",
   "Single export": "တစ်ခုခြင်း တင်ပို့ရန်",
   "SMS": "စာတို",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "ခိုင်ခန့်သော ဘလော့ခ်အက္ခရာများ (██) — ခေတ်မီ ဖောင့်များတွင် အကောင်းဆုံး တိကျမှန်ကန်မှု။",
   "Sponsor": "စပွန်ဆာ",
   "square": "စတုရန်း",
   "Start Date": "စတင်ရက်",
@@ -291,15 +291,15 @@
   "Text": "စာသား",
   "Text color": "စာသား အရောင်",
   "Text is required": "စာသား လိုအပ်သည်",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "စာသားထုတ်လွှင့်မှုတွင် ဖရိမ်နှင့် ပုံ မပါဝင်ပါ။",
   "Text position": "စာသား နေရာ",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "ကလစ်ဘုတ်သို့ တစ်ကြိမ်တည်းဖြင့် ကူးယူရန် ကျူအာကုဒ်များ များပြားလွန်းနေသည်။ ယင်းတို့အား SVG သို့ PNG ဖိုင်များအဖြစ် ကျေးဇူးပြု၍ ဒေါင်းလုပ် လုပ်ပါ။",
   "Tip: For best results, use a clear image with good lighting.": "အကြံပြုချက် - အကောင်းဆုံးရလဒ်များအတွက် အလင်းရောင်ကောင်းမွန်ပြီး ရှင်းလင်းသောပုံကို အသုံးပြုပါ။",
   "Title": "ခေါင်းစဉ်",
   "Toggle dark mode": "အမှောင်မုဒ် ချိန်ညှိရန်",
   "top": "ထိပ်ဆုံး",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "ယူနီကုဒ် (အပြည့်ဘလော့ခ်များ)",
+  "Unicode (half-blocks)": "ယူနီကုဒ် (ဝက်ဘလော့ခ်များ)",
   "Upload a CSV file": "CSV ဖိုင်တစ်ခုအား တင်ရန်",
   "Upload image": "ပုံတင်ရန်",
   "Upload QR Code Image": "ကျူအာကုဒ်ပုံအား တင်ရန်",
