+++ locales/nl.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Geladen vanuit lokale opslag",
   "{count} piece(s) of data detected": "{count} stuk(en) van data gedetecteerd",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR codes zijn aangemaakt.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bits tekens — werkt in platte tekst, e-mail en terminals.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bits, omgekeerde polariteit (# is achtergrond) — werkt in platte tekst, e-mail en terminals.",
   "Add frame": "Frame toevoegen",
   "Address": "Adres",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Dialoog sluiten",
   "Close guide": "Sluit gids",
   "Close scanner": "Sluit scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: twee rijen samengevoegd tot één regel met behulp van de tekens ▀ ▄ █.",
   "Configure frame settings": "Frame-instellingen configureren",
   "Contact Card": "Contact kaart",
-  "Copied": "Copied",
+  "Copied": "Gekopieerd",
   "Copy QR Code to clipboard": "QR-Code kopiëren naar klembord",
   "Copy to clipboard": "Kopiëren naar klembord",
   "Corners Dot color": "Puntkleur van de hoeken",
@@ -113,10 +113,10 @@
   "dots": "stippen",
   "Dots color": "Kleur stippen",
   "Dots type": "Soort stippen",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".md downloaden",
+  "Download .txt": ".txt downloaden",
+  "Download all as .md": "Alles downloaden als .md",
+  "Download all as .txt": "Alles downloaden als .txt",
   "Download QR Code as JPG": "Download QR-Code als JPG",
   "Download QR Code as PNG": "Download QR-Code als PNG",
   "Download QR Code as SVG": "Download QR-Code als SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Voorbeeldbestand (vCard)",
   "Example with Custom Filenames": "Voorbeeld met aangepaste bestandsnamen",
   "Export": "Exporteren",
-  "Export as text": "Export as text",
+  "Export as text": "Exporteren als tekst",
   "Export QR code": "QR-code exporteren",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR-code exporteren als ASCII- of Unicode-tekst",
   "extra-rounded": "extra-afgerond",
   "Failed to copy to clipboard": "Kon niet naar klembord kopiëren",
   "Failed to load changelog": "Changelog niet geladen",
@@ -161,12 +161,12 @@
   "First Name": "Voornaam",
   "First row preview:": "Voorvertoning eerste rij:",
   "Font family": "Lettertypefamilie",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Het framelabel wordt niet meegenomen bij de tekstexport — de ASCII-uitvoer bevat alleen de QR-modules.",
   "Frame preset": "Vooringesteld frame",
   "Frame settings": "Frame instellingen",
   "Frame style": "Frame stijl",
   "Frame text": "Frame tekst",
-  "Generating": "Generating",
+  "Generating": "Genereren",
   "GitHub repository for this project": "GitHub repository voor dit project",
   "Have a beautiful day!": "Heb een mooie dag!",
   "Have nice day!": "Fijne dag!",
@@ -204,7 +204,7 @@
   "Message": "Bericht",
   "Mini QR Code Generator": "Mini QR-codegenerator",
   "MiniQR's QR rendering engine has been updated": "De QR-weergave-engine van MiniQR is bijgewerkt",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "De meeste QR-scanners kunnen dit formaat niet lezen omdat de #-tekens tussen de rijen ruimte laten. Gebruik Unicode (halve/volledige blokken) voor een scanbare export.",
   "Must be between 0 and 1": "Moet tussen 0 en 1 liggen",
   "Network Name (SSID)": "Netwerknaam (SSID)",
   "Network password (case-sensitive)": "Netwerkwachtwoord (hoofdlettergevoelig)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Upload eerst een bestand",
   "Position": "Positie",
   "Preset": "Vooraf ingesteld",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Voorbeeld van de eerste rij",
   "Processing...": "Verwerken...",
   "QR Code Content": "Inhoud QR-code",
   "QR code preset": "QR-code vooraf ingesteld",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Eenvoudige URL/tekst",
   "Single export": "Enkele export",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Volledige bloktekens (██) — de beste weergave in moderne lettertypen.",
   "Sponsor": "Sponsor",
   "square": "Vierkant",
   "Start Date": "Startdatum",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Tekst kleur",
   "Text is required": "Tekst is vereist",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "De tekstuitvoer omvat niet het kader en de afbeelding",
   "Text position": "Tekst positie",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Er zijn te veel QR-codes om tegelijk naar het klembord te worden gekopieerd. Download ze als SVG of PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Tip: Voor de beste resultaten gebruik je een duidelijke afbeelding met een goede verlichting.",
   "Title": "Titel",
   "Toggle dark mode": "Donkere modus in-/uitschakelen",
   "top": "bovenkant",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (volledige blokken)",
+  "Unicode (half-blocks)": "Unicode (halve blokken)",
   "Upload a CSV file": "CSV-bestand uploaden",
   "Upload image": "Afbeelding uploaden",
   "Upload QR Code Image": "Upload QR Code afbeelding",
