+++ locales/pl.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Załadowano z pamięci lokalnej",
   "{count} piece(s) of data detected": "Wykryto {count} fragment(y) danych",
   "{index} / {count} QR codes have been created.": "{index} / {count} kody QR zostały utworzone.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Znaki 7-bitowe — działają w tekście zwykłym, wiadomościach e-mail i terminalach.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bitowy, odwrócona polaryzacja (znak # to tło) — działa w tekście zwykłym, wiadomościach e-mail i terminalach.",
   "Add frame": "Dodaj ramkę",
   "Address": "Adres",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Zamknij okno dialogowe",
   "Close guide": "Zamknij przewodnik",
   "Close scanner": "Zamknij skaner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktowy: dwa rzędy ułożone w jedną linię przy użyciu znaków ▀ ▄ █.",
   "Configure frame settings": "Konfiguracja ustawień ramki",
   "Contact Card": "Karta kontaktowa",
-  "Copied": "Copied",
+  "Copied": "Skopiowano",
   "Copy QR Code to clipboard": "Skopiuj kod QR do schowka",
   "Copy to clipboard": "Kopiuj do schowka",
   "Corners Dot color": "Kolor kropki narożników",
@@ -113,10 +113,10 @@
   "dots": "kropki",
   "Dots color": "Kolor kropek",
   "Dots type": "Typ kropek",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Pobierz plik .md",
+  "Download .txt": "Pobierz plik .txt",
+  "Download all as .md": "Pobierz wszystko w formacie .md",
+  "Download all as .txt": "Pobierz wszystko jako plik .txt",
   "Download QR Code as JPG": "Pobierz kod QR jako JPG",
   "Download QR Code as PNG": "Pobierz kod QR jako PNG",
   "Download QR Code as SVG": "Pobierz kod QR jako SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Przykładowy plik (vCard)",
   "Example with Custom Filenames": "Przykład z niestandardowymi nazwami plików",
   "Export": "Eksportuj",
-  "Export as text": "Export as text",
+  "Export as text": "Eksportuj jako tekst",
   "Export QR code": "Eksportuj kod QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksportuj kod QR jako tekst w formacie ASCII lub Unicode",
   "extra-rounded": "ekstrazaokrąglony",
   "Failed to copy to clipboard": "Nie udało się skopiować do schowka",
   "Failed to load changelog": "Nie udało się załadować dziennika zmian",
@@ -161,12 +161,12 @@
   "First Name": "Imię",
   "First row preview:": "Podgląd pierwszego wiersza:",
   "Font family": "Rodzina czcionek",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Etykieta ramki nie jest uwzględniana w eksporcie tekstowym — wynik w formacie ASCII zawiera wyłącznie moduły QR.",
   "Frame preset": "Wstępne ustawienie ramki",
   "Frame settings": "Ustawienia ramki",
   "Frame style": "Styl ramki",
   "Frame text": "Tekst ramki",
-  "Generating": "Generating",
+  "Generating": "Generowanie",
   "GitHub repository for this project": "Repozytorium GitHub dla tego projektu",
   "Have a beautiful day!": "Miłego dnia!",
   "Have nice day!": "Miłego dnia!",
@@ -204,7 +204,7 @@
   "Message": "Wiadomość",
   "Mini QR Code Generator": "Mini generator kodów QR",
   "MiniQR's QR rendering engine has been updated": "Silnik renderowania kodów QR w aplikacji MiniQR został zaktualizowany",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Większość skanerów kodów QR nie odczyta tego formatu, ponieważ znaki # powodują powstanie odstępów między wierszami. Aby uzyskać plik nadający się do skanowania, należy użyć znaków Unicode (bloki półpełne lub pełne).",
   "Must be between 0 and 1": "Wartość musi wynosić od 0 do 1",
   "Network Name (SSID)": "Nazwa sieci (SSID)",
   "Network password (case-sensitive)": "Hasło sieciowe (wielkość liter ma znaczenie)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Proszę najpierw przesłać plik",
   "Position": "Położenie",
   "Preset": "Ustawienia wstępne",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Podgląd pierwszego wiersza",
   "Processing...": "Przetwarzanie...",
   "QR Code Content": "Zawartość kodu QR",
   "QR code preset": "Ustawienia kodu QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Prosty adres URL/tekst",
   "Single export": "Pojedynczy eksport",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Znaki w postaci jednolitych prostokątów (██) — zapewniają najwyższą wierność odwzorowania w nowoczesnych czcionkach.",
   "Sponsor": "Jednostka sponsorująca",
   "square": "kwadrat",
   "Start Date": "Data rozpoczęcia",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Kolor tekstu",
   "Text is required": "Wymagany jest tekst",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Wynik tekstowy nie obejmuje ramki ani obrazu",
   "Text position": "Pozycja tekstu",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Istnieje zbyt wiele kodów QR do skopiowania do schowka jednocześnie. Pobierz je jako SVG lub PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Wskazówka: Aby uzyskać najlepsze wyniki, użyj czystego obrazu z dobrym oświetleniem.",
   "Title": "Tytuł",
   "Toggle dark mode": "Przełącz tryb ciemny",
   "top": "góra",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (bloki pełne)",
+  "Unicode (half-blocks)": "Unicode (półbloki)",
   "Upload a CSV file": "Prześlij plik CSV",
   "Upload image": "Prześlij obraz",
   "Upload QR Code Image": "Prześlij obraz kodu QR",
