+++ locales/pt.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Carregado do armazenamento local",
   "{count} piece(s) of data detected": "{count} dados detetados",
   "{index} / {count} QR codes have been created.": "{index} / {count} Foram criados códigos QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Caracteres de 7 bits — funcionam em texto simples, e-mail e terminais.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bits, polaridade invertida (# é o fundo) — funciona em texto simples, e-mail e terminais.",
   "Add frame": "Adicionar moldura",
   "Address": "Endereço",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Fechar diálogo",
   "Close guide": "Fechar guia",
   "Close scanner": "Fechar scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compacto: duas linhas agrupadas em uma única linha usando os caracteres ▀ ▄ █.",
   "Configure frame settings": "Configurar definições do quadro",
   "Contact Card": "Cartão de contacto",
-  "Copied": "Copied",
+  "Copied": "Copiado",
   "Copy QR Code to clipboard": "Copiar código QR para a área de transferência",
   "Copy to clipboard": "Copiar para a área de transferência",
   "Corners Dot color": "Cantos Cor dos pontos",
@@ -113,10 +113,10 @@
   "dots": "pontos",
   "Dots color": "Cor dos pontos",
   "Dots type": "Tipo de pontos",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Baixar .md",
+  "Download .txt": "Baixar .txt",
+  "Download all as .md": "Baixar tudo como .md",
+  "Download all as .txt": "Baixar tudo como .txt",
   "Download QR Code as JPG": "Descarregar código QR como JPG",
   "Download QR Code as PNG": "Descarregar código QR como PNG",
   "Download QR Code as SVG": "Descarregar código QR como SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Ficheiro de exemplo (vCard)",
   "Example with Custom Filenames": "Exemplo com nomes de ficheiros personalizados",
   "Export": "Exportar",
-  "Export as text": "Export as text",
+  "Export as text": "Exportar como texto",
   "Export QR code": "Exportar código QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportar o código QR como texto ASCII ou Unicode",
   "extra-rounded": "extra arredondado",
   "Failed to copy to clipboard": "Falha ao copiar para a área de transferência",
   "Failed to load changelog": "Falha ao carregar o registo de alterações",
@@ -161,12 +161,12 @@
   "First Name": "Nome",
   "First row preview:": "Pré-visualização da primeira linha:",
   "Font family": "Família de fontes",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "A etiqueta do quadro é excluída da exportação de texto — a saída em ASCII contém apenas os módulos QR.",
   "Frame preset": "Predefinição do quadro",
   "Frame settings": "Configurações do quadro",
   "Frame style": "Estilo da moldura",
   "Frame text": "Texto do quadro",
-  "Generating": "Generating",
+  "Generating": "Gerando",
   "GitHub repository for this project": "Repositório GitHub para este projeto",
   "Have a beautiful day!": "Tenha um ótimo dia!",
   "Have nice day!": "Tenha um bom dia!",
@@ -204,7 +204,7 @@
   "Message": "Mensagem",
   "Mini QR Code Generator": "Mini Gerador de QR Code",
   "MiniQR's QR rendering engine has been updated": "O mecanismo de renderização de QR codes do MiniQR foi atualizado",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "A maioria dos leitores de QR Code não consegue ler esse formato porque os caracteres # deixam espaços entre as linhas. Use Unicode (blocos de meio/blocos inteiros) para uma exportação que possa ser lida por leitores de QR Code.",
   "Must be between 0 and 1": "Deve estar entre 0 e 1",
   "Network Name (SSID)": "Nome da rede (SSID)",
   "Network password (case-sensitive)": "Senha da rede (sensível a maiúsculas e minúsculas)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Por favor, carregue primeiro um ficheiro",
   "Position": "Posição",
   "Preset": "Predefinição",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Pré-visualização da primeira linha",
   "Processing...": "A processar...",
   "QR Code Content": "Conteúdo do código QR",
   "QR code preset": "Código QR predefinido",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/texto simples",
   "Single export": "Exportação única",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Caracteres em blocos sólidos (██) — a melhor fidelidade em fontes modernas.",
   "Sponsor": "Patrocinador",
   "square": "quadrado",
   "Start Date": "Data de início",
@@ -291,15 +291,15 @@
   "Text": "Texto",
   "Text color": "Cor do texto",
   "Text is required": "É necessário inserir texto",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "A saída de texto exclui a moldura e a imagem",
   "Text position": "Posição do texto",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Existem demasiados códigos QR para serem copiados para a área de transferência de uma só vez. Por favor, descarregue-os como SVG ou PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Dica: Para obter melhores resultados, use uma imagem nítida com boa iluminação.",
   "Title": "Título",
   "Toggle dark mode": "Alternar modo escuro",
   "top": "topo",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blocos completos)",
+  "Unicode (half-blocks)": "Unicode (meios blocos)",
   "Upload a CSV file": "Carregar um ficheiro CSV",
   "Upload image": "Carregar imagem",
   "Upload QR Code Image": "Carregar imagem do código QR",
