+++ locales/ro.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Încărcat din stocarea locală",
   "{count} piece(s) of data detected": "{count} bucăți de date detectate",
   "{index} / {count} QR codes have been created.": "{index} / {count} au fost create coduri QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Caractere pe 7 biți — funcționează în text simplu, e-mail și terminale.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 biți, polaritate inversată (# reprezintă fundalul) — funcționează în text simplu, e-mail și terminale.",
   "Add frame": "Adaugă cadru",
   "Address": "Adresă",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Închideți dialogul",
   "Close guide": "Închide ghidul",
   "Close scanner": "Închide scanerul",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: două rânduri înghesuite într-un singur rând folosind caracterele ▀ ▄ █.",
   "Configure frame settings": "Configurați setările cadrului",
   "Contact Card": "Card de contact",
-  "Copied": "Copied",
+  "Copied": "Copiat",
   "Copy QR Code to clipboard": "Copiază codul QR în clipboard",
   "Copy to clipboard": "Copiază în clipboard",
   "Corners Dot color": "Colț punct culoare",
@@ -113,10 +113,10 @@
   "dots": "puncte",
   "Dots color": "Puncte de culoare",
   "Dots type": "Tip puncte",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Descărcați fișierul .md",
+  "Download .txt": "Descărcați fișierul .txt",
+  "Download all as .md": "Descărcați totul în format .md",
+  "Download all as .txt": "Descărcați totul în format .txt",
   "Download QR Code as JPG": "Descarcă codul QR ca JPG",
   "Download QR Code as PNG": "Descarcă codul QR ca PNG",
   "Download QR Code as SVG": "Descarcă codul QR ca SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Exemplu de fișier (vCard)",
   "Example with Custom Filenames": "Exemplu cu nume de fișiere personalizate",
   "Export": "Exportă",
-  "Export as text": "Export as text",
+  "Export as text": "Exportă ca text",
   "Export QR code": "Exportă codul QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportă codul QR ca text ASCII sau Unicode",
   "extra-rounded": "extra rotunjit",
   "Failed to copy to clipboard": "Copierea în clipboard a eșuat",
   "Failed to load changelog": "A eșuat încărcarea jurnalului de schimbări",
@@ -161,12 +161,12 @@
   "First Name": "Numele și prenumele",
   "First row preview:": "Previzualizare primul rând:",
   "Font family": "Familia de fonturi",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Eticheta cadrului este exclusă din exportul de text — fișierul ASCII conține doar modulele QR.",
   "Frame preset": "Cadru prestabilit",
   "Frame settings": "Setări cadru",
   "Frame style": "Stil cadru",
   "Frame text": "Text cadru",
-  "Generating": "Generating",
+  "Generating": "Generarea",
   "GitHub repository for this project": "Depozitul GitHub pentru acest proiect",
   "Have a beautiful day!": "Să aveți o zi frumoasă!",
   "Have nice day!": "O zi frumoasă!",
@@ -204,7 +204,7 @@
   "Message": "Mesaj",
   "Mini QR Code Generator": "Generator mini cod QR",
   "MiniQR's QR rendering engine has been updated": "Motorul de redare QR al MiniQR a fost actualizat",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Majoritatea scanerelor QR nu vor citi acest format, deoarece simbolurile # lasă spații între rânduri. Folosiți Unicode (blocuri jumătate/întregi) pentru un export care poate fi scanat.",
   "Must be between 0 and 1": "Trebuie să fie între 0 și 1",
   "Network Name (SSID)": "Numele rețelei (SSID)",
   "Network password (case-sensitive)": "Parola rețelei (sensibilă la majuscule)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Vă rugăm să încărcaţi mai întâi un fişier",
   "Position": "Poziție",
   "Preset": "Presetare",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Previzualizare rândul 1",
   "Processing...": "Procesare...",
   "QR Code Content": "Conținut cod QR",
   "QR code preset": "Cod QR prestabilit",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL simplu/Text",
   "Single export": "Export unic",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Caractere sub formă de blocuri pline (██) — cea mai bună redare în fonturile moderne.",
   "Sponsor": "Sponsor",
   "square": "pătrat",
   "Start Date": "Data de începere",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Culoare text",
   "Text is required": "Textul este necesar",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Textul afișat nu include cadrul și imaginea",
   "Text position": "Poziție text",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Există prea multe coduri QR pentru a fi copiate în clipboard simultan. Vă rugăm să le descărcați ca SVG sau PNG în schimb.",
   "Tip: For best results, use a clear image with good lighting.": "Sfat: Pentru cele mai bune rezultate, folosește o imagine clară cu lumină bună.",
   "Title": "Titlu",
   "Toggle dark mode": "Comută modul întunecat",
   "top": "sus",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blocuri complete)",
+  "Unicode (half-blocks)": "Unicode (semiblocuri)",
   "Upload a CSV file": "Încărcați un fișier CSV",
   "Upload image": "Încarcă imaginea",
   "Upload QR Code Image": "Încărcați imaginea codului QR",
