+++ locales/ru.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Загружено из локального хранилища",
   "{count} piece(s) of data detected": "{count} часть(и) данных обнаружено",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-коды созданы.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-битные символы — поддерживаются в простом тексте, электронной почте и терминалах.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-битный, с инвертированной полярностью (# — фон) — поддерживается в текстовых файлах, электронной почте и терминалах.",
   "Add frame": "Добавить кадр",
   "Address": "Адрес",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Закрыть диалог",
   "Close guide": "Закрыть руководство",
   "Close scanner": "Закрыть сканер",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Компактный вариант: два ряда, сгруппированные в одну строку с помощью символов ▀ ▄ █.",
   "Configure frame settings": "Настройка параметров кадра",
   "Contact Card": "Контактная карта",
-  "Copied": "Copied",
+  "Copied": "Скопировано",
   "Copy QR Code to clipboard": "Скопировать QR-код в буфер обмена",
   "Copy to clipboard": "Копировать в буфер обмена",
   "Corners Dot color": "Цвет точки углов",
@@ -113,10 +113,10 @@
   "dots": "точки",
   "Dots color": "Цвет точек",
   "Dots type": "Тип точек",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Скачать .md",
+  "Download .txt": "Скачать файл .txt",
+  "Download all as .md": "Скачать всё в формате .md",
+  "Download all as .txt": "Скачать всё в формате .txt",
   "Download QR Code as JPG": "Скачать QR-код как JPG",
   "Download QR Code as PNG": "Скачать QR-код как PNG",
   "Download QR Code as SVG": "Скачать QR-код как SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Пример файла (vCard)",
   "Example with Custom Filenames": "Пример с настраиваемыми именами файлов",
   "Export": "Экспорт",
-  "Export as text": "Export as text",
+  "Export as text": "Экспорт в текстовый формат",
   "Export QR code": "Экспорт QR-кода",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Экспорт QR-кода в виде текста в формате ASCII или Unicode",
   "extra-rounded": "неокруглённый",
   "Failed to copy to clipboard": "Не удалось скопировать в буфер обмена",
   "Failed to load changelog": "Не удалось загрузить журнал изменений",
@@ -161,12 +161,12 @@
   "First Name": "Имя",
   "First row preview:": "Предпросмотр первой строки:",
   "Font family": "Семейство шрифтов",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Этикетка кадра не включается в экспорт текста — в ASCII-файле содержатся только модули QR-кода.",
   "Frame preset": "Предварительная настройка кадра",
   "Frame settings": "Настройки кадров",
   "Frame style": "Стиль кадра",
   "Frame text": "Текст рамки",
-  "Generating": "Generating",
+  "Generating": "Генерация",
   "GitHub repository for this project": "GitHub репозиторий для этого проекта",
   "Have a beautiful day!": "Хорошего дня!",
   "Have nice day!": "Приятного дня!",
@@ -204,7 +204,7 @@
   "Message": "Сообщение",
   "Mini QR Code Generator": "Генератор мини-QR-кодов",
   "MiniQR's QR rendering engine has been updated": "Обновлен движок рендеринга QR-кодов MiniQR",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Большинство QR-сканеров не смогут распознать этот формат, поскольку символы # оставляют пробелы между строками. Для экспорта в сканируемом формате используйте Unicode (полублоки/полные блоки).",
   "Must be between 0 and 1": "Должно быть в диапазоне от 0 до 1",
   "Network Name (SSID)": "Имя сети (SSID)",
   "Network password (case-sensitive)": "Сетевой пароль (с учетом регистра)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Сначала загрузите файл",
   "Position": "Позиция",
   "Preset": "Предустановка",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Предварительный просмотр первой строки",
   "Processing...": "Обработка...",
   "QR Code Content": "Содержание QR-кода",
   "QR code preset": "Предустановленный QR-код",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Простой URL/текст",
   "Single export": "Одиночный экспорт",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Сплошные символы (██) — наилучшее качество отображения в современных шрифтах.",
   "Sponsor": "Спонсор",
   "square": "квадрат",
   "Start Date": "Дата начала",
@@ -291,15 +291,15 @@
   "Text": "Текст",
   "Text color": "Цвет текста",
   "Text is required": "Текст обязателен",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "В текстовом выводе не отображаются рамка и изображение",
   "Text position": "Позиция текста",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Слишком много QR-кодов для копирования в буфер обмена сразу. Пожалуйста, скачайте их как SVG или PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Подсказка: Для лучшего результата используйте четкое изображение с хорошим освещением.",
   "Title": "Название",
   "Toggle dark mode": "Переключить темный режим",
   "top": "сверху",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (полные блоки)",
+  "Unicode (half-blocks)": "Unicode (полублоки)",
   "Upload a CSV file": "Загрузите файл CSV",
   "Upload image": "Загрузить изображение",
   "Upload QR Code Image": "Загрузить изображение QR-кода",
