+++ locales/sk.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Načítané z miestneho úložiska",
   "{count} piece(s) of data detected": "{count} zistené údaje",
   "{index} / {count} QR codes have been created.": "{index} / {count} Boli vytvorené QR kódy.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bitové znaky – fungujú v prostom texte, e-mailoch a termináloch.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bitové, s obrátenou polaritou (znak # je pozadie) — funguje v prostom texte, e-mailoch a termináloch.",
   "Add frame": "Pridať rám",
   "Address": "Adresa",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Zavrieť dialógové okno",
   "Close guide": "Zatvoriť sprievodcu",
   "Close scanner": "Zavrieť skener",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompaktné: dva riadky zjednotené do jedného pomocou znakov ▀ ▄ █.",
   "Configure frame settings": "Konfigurácia nastavení rámu",
   "Contact Card": "Kontaktná karta",
-  "Copied": "Copied",
+  "Copied": "Skopírované",
   "Copy QR Code to clipboard": "Skopírovanie kódu QR do schránky",
   "Copy to clipboard": "Kopírovať do schránky",
   "Corners Dot color": "Rohy Farba bodky",
@@ -113,10 +113,10 @@
   "dots": "bodky",
   "Dots color": "Farba bodiek",
   "Dots type": "Typ bodiek",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Stiahnuť .md",
+  "Download .txt": "Stiahnuť súbor .txt",
+  "Download all as .md": "Stiahnuť všetko ako .md",
+  "Download all as .txt": "Stiahnuť všetko ako súbor .txt",
   "Download QR Code as JPG": "Stiahnutie kódu QR ako JPG",
   "Download QR Code as PNG": "Stiahnutie kódu QR ako PNG",
   "Download QR Code as SVG": "Stiahnutie kódu QR ako SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Príklad súboru (vCard)",
   "Example with Custom Filenames": "Príklad s vlastnými názvami súborov",
   "Export": "Export",
-  "Export as text": "Export as text",
+  "Export as text": "Exportovať ako text",
   "Export QR code": "Exportovať QR kód",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportovať QR kód ako text vo formáte ASCII alebo Unicode",
   "extra-rounded": "mimoriadne zaoblené",
   "Failed to copy to clipboard": "Nepodarilo sa skopírovať do schránky",
   "Failed to load changelog": "Nepodarilo sa načítať zoznam zmien",
@@ -161,12 +161,12 @@
   "First Name": "Kresťanské meno",
   "First row preview:": "Náhľad do prvého radu:",
   "Font family": "Rodina písem",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Štítok rámu nie je súčasťou exportu textu – výstup v kódovaní ASCII obsahuje iba moduly QR.",
   "Frame preset": "Prednastavenie rámu",
   "Frame settings": "Nastavenia rámu",
   "Frame style": "Štýl rámu",
   "Frame text": "Text rámu",
-  "Generating": "Generating",
+  "Generating": "Generovanie",
   "GitHub repository for this project": "Úložisko GitHub pre tento projekt",
   "Have a beautiful day!": "Prajem vám krásny deň!",
   "Have nice day!": "Prajem pekný deň!",
@@ -204,7 +204,7 @@
   "Message": "Správa",
   "Mini QR Code Generator": "Generátor mini QR kódov",
   "MiniQR's QR rendering engine has been updated": "Bol aktualizovaný engine na generovanie QR kódov v aplikácii MiniQR",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Väčšina čítačiek QR kódov tento formát neprečíta, pretože znaky # vytvárajú medzery medzi riadkami. Ak chcete exportovať formát, ktorý je možné naskenovať, použite Unicode (polovičné/celé bloky).",
   "Must be between 0 and 1": "Musí byť v rozmedzí od 0 do 1",
   "Network Name (SSID)": "Názov siete (SSID)",
   "Network password (case-sensitive)": "Sieťové heslo (rozlišuje sa podľa veľkosti písmen)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Najprv nahrajte súbor",
   "Position": "Pozícia",
   "Preset": "Prednastavené",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Náhľad prvého riadku",
   "Processing...": "Spracovanie...",
   "QR Code Content": "Obsah kódu QR",
   "QR code preset": "Predvoľba kódu QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Jednoduchá URL adresa/text",
   "Single export": "Jediný vývoz",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Plné blokové znaky (██) — najvyššia vernosť v moderných písmach.",
   "Sponsor": "Sponzor",
   "square": "štvorcový",
   "Start Date": "Dátum začiatku",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Farba textu",
   "Text is required": "Vyžaduje sa text",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Výstupný text nezahŕňa rámček ani obrázok",
   "Text position": "Pozícia textu",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Do schránky sa naraz skopíruje príliš veľa kódov QR. Namiesto toho si ich stiahnite ako SVG alebo PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Tip: Pre dosiahnutie najlepších výsledkov použite jasný obraz s dobrým osvetlením.",
   "Title": "Názov",
   "Toggle dark mode": "Prepnúť tmavý režim",
   "top": "vrchol",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (celé bloky)",
+  "Unicode (half-blocks)": "Unicode (polobloky)",
   "Upload a CSV file": "Nahratie súboru CSV",
   "Upload image": "Nahrať obrázok",
   "Upload QR Code Image": "Nahratie obrázka kódu QR",
