+++ locales/sq.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Ngarkuar nga ruajtja lokale",
   "{count} piece(s) of data detected": "{count} Copë(a) të dhënash të zbuluar",
   "{index} / {count} QR codes have been created.": "{index} / {count} Janë krijuar kode QR.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Karaktet 7-bit — funksionojnë në tekst të thjeshtë, email dhe terminale.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, polaritet i kundërt (# është sfondi) — funksionon në tekst të thjeshtë, email dhe terminale.",
   "Add frame": "Shto kornizë",
   "Address": "Adresa",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Mbyll dialogun",
   "Close guide": "Mbyll udhëzuesin",
   "Close scanner": "Mbyll skanerën",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompakt: dy rreshta të ngjeshura në një linjë duke përdorur karakteret ▀ ▄ █.",
   "Configure frame settings": "Konfiguro cilësimet e kornizës",
   "Contact Card": "Kartë kontakti",
-  "Copied": "Copied",
+  "Copied": "Kopjuar",
   "Copy QR Code to clipboard": "Kopjo kodin QR në clipboard",
   "Copy to clipboard": "Kopjo në clipboard",
   "Corners Dot color": "Këndet pika Dot ngjyra",
@@ -113,10 +113,10 @@
   "dots": "pikëpikë",
   "Dots color": "Ngjyros pikat",
   "Dots type": "Lloji i pikave",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Shkarko .md",
+  "Download .txt": "Shkarko .txt",
+  "Download all as .md": "Shkarko të gjitha si .md",
+  "Download all as .txt": "Shkarko të gjitha si .txt",
   "Download QR Code as JPG": "Shkarko Kodin QR si JPG",
   "Download QR Code as PNG": "Shkarko Kodin QR si PNG",
   "Download QR Code as SVG": "Shkarko Kodin QR si SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Skedar shembull (vCard)",
   "Example with Custom Filenames": "Shembull me emra skedarësh të personalizuar",
   "Export": "Eksport",
-  "Export as text": "Export as text",
+  "Export as text": "Eksportoni si tekst",
   "Export QR code": "Eksporto kodin QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Eksportoni kodin QR si tekst ASCII ose Unicode",
   "extra-rounded": "shumë i rrumbullakët",
   "Failed to copy to clipboard": "Dështoi kopjimi në clipboard",
   "Failed to load changelog": "Dështoi ngarkimi i listës së ndryshimeve",
@@ -161,12 +161,12 @@
   "First Name": "Emri i parë",
   "First row preview:": "Parathënie e rreshtit të parë:",
   "Font family": "Familja e fontit",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Etiketa e kornizës përjashtohet nga eksporti i tekstit — dalja ASCII përmban vetëm modulet QR.",
   "Frame preset": "Paracaktim kornize",
   "Frame settings": "Caktimet e kornizës",
   "Frame style": "Stili i kornizës",
   "Frame text": "Teksti i kornizës",
-  "Generating": "Generating",
+  "Generating": "Gjenerim",
   "GitHub repository for this project": "Depozita GitHub për këtë projekt",
   "Have a beautiful day!": "Të kesh një ditë të bukur!",
   "Have nice day!": "Të kesh një ditë të mirë!",
@@ -204,7 +204,7 @@
   "Message": "Mesazh",
   "Mini QR Code Generator": "Gjenerator i vogël i kodeve QR",
   "MiniQR's QR rendering engine has been updated": "Motori i renderimit të QR-ve të MiniQR-it është përditësuar.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Shumica e skanerëve QR nuk do ta lexojnë këtë format sepse glifet # lënë boshllëqe midis rreshtave. Përdorni Unicode (gjysmë/blloqe të plota) për një eksport të skanueshëm.",
   "Must be between 0 and 1": "Duhet të jetë midis 0 dhe 1",
   "Network Name (SSID)": "Emri i rrjetit (SSID)",
   "Network password (case-sensitive)": "Fjalëkalimi i rrjetit (i ndjeshëm ndaj shkronjave të mëdha dhe të vogla)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Ju lutem ngarkoni së pari një skedar.",
   "Position": "Pozicion",
   "Preset": "Paracaktim",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Parathënie e rreshtit të parë",
   "Processing...": "Përpunim...",
   "QR Code Content": "Përmbajtja e Kodit QR",
   "QR code preset": "Paracaktim i kodit QR",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/Tekst i thjeshtë",
   "Single export": "Eksport i vetëm",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Karakteret bllok të ngurta (██) — besueshmëria më e mirë në fontet moderne.",
   "Sponsor": "Sponsor",
   "square": "katrore",
   "Start Date": "Data e fillimit",
@@ -291,15 +291,15 @@
   "Text": "Tekst",
   "Text color": "Ngjyra e tekstit",
   "Text is required": "Teksti është i nevojshëm",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Dalja e tekstit përjashton kornizën dhe imazhin.",
   "Text position": "Pozicioni i tekstit",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Ka shumë kode QR që mund të kopjohen në clipboard njëherësh. Ju lutemi shkarkojini ato si SVG ose PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Këshillë: Për rezultatet më të mira, përdorni një imazh të qartë me ndriçim të mirë.",
   "Title": "Titulli",
   "Toggle dark mode": "Alternoni modalitetin e errët",
   "top": "majë",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (blloqe të plota)",
+  "Unicode (half-blocks)": "Unikod (gjysmë-blloqe)",
   "Upload a CSV file": "Ngarko një skedar CSV",
   "Upload image": "Ngarko imazhin",
   "Upload QR Code Image": "Ngarko imazhin e kodit QR",
