+++ locales/sr.json
@@ -51,11 +51,11 @@
   "Last saved locally": "Учитано из локалног складишта",
   "{count} piece(s) of data detected": "{count} делови података откривени",
   "{index} / {count} QR codes have been created.": "{index} / {count} Кодни QR кодови су креирани.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-битни карактери — ради у обичном тексту, електронској пошти и терминалима.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-битни, инверзна поларитет (# је позадина) — ради у обичном тексту, е-пошти и терминалима.",
   "Add frame": "Додај оквир",
   "Address": "Адреса",
-  "ASCII": "ASCII",
+  "ASCII": "АСЦИИ",
   "Available Fields": "Доступна поља",
   "Background color": "Боја позадине",
   "Batch export": "Масовни извоз",
@@ -82,10 +82,10 @@
   "Close dialog": "Затвори дијалог",
   "Close guide": "Затвори водич",
   "Close scanner": "Затвори скенер",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Компактно: два реда спакована у један низ коришћењем знакова ▀ ▄ █.",
   "Configure frame settings": "Конфигуришите подешавања оквира",
   "Contact Card": "Визит-карта",
-  "Copied": "Copied",
+  "Copied": "Копирано",
   "Copy QR Code to clipboard": "Копирајте QR код у међуспремник",
   "Copy to clipboard": "Копирај у међуспремник",
   "Corners Dot color": "Углови Дота боја",
@@ -113,10 +113,10 @@
   "dots": "тачке",
   "Dots color": "Боја тачака",
   "Dots type": "Тип тачака",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Преузми .md",
+  "Download .txt": "Преузми .txt",
+  "Download all as .md": "Преузмите све као .md",
+  "Download all as .txt": "Преузмите све као .txt",
   "Download QR Code as JPG": "Преузмите QR код као JPG",
   "Download QR Code as PNG": "Преузмите QR код као PNG",
   "Download QR Code as SVG": "Преузми QR код као SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Пример фајла (vCard)",
   "Example with Custom Filenames": "Пример са прилагођеним именима датотека",
   "Export": "Извоз",
-  "Export as text": "Export as text",
+  "Export as text": "Извези као текст",
   "Export QR code": "Извези QR код",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Извези QR код као ASCII или Unicode текст",
   "extra-rounded": "посебно заобљен",
   "Failed to copy to clipboard": "Копирање у међуспречник није успело",
   "Failed to load changelog": "Учитавање лога промена није успело",
@@ -161,12 +161,12 @@
   "First Name": "Прво име",
   "First row preview:": "Преглед првог реда:",
   "Font family": "Породица фонтова",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Ознака оквира није укључена у извоз текста — ASCII излаз садржи само QR модуле.",
   "Frame preset": "Подешавање оквира",
   "Frame settings": "Подешавања оквира",
   "Frame style": "Стил оквира",
   "Frame text": "Текст оквира",
-  "Generating": "Generating",
+  "Generating": "Генерисање",
   "GitHub repository for this project": "GitHub репозиторијум за овај пројекат",
   "Have a beautiful day!": "Имајте леп дан!",
   "Have nice day!": "Имајте леп дан!",
@@ -204,7 +204,7 @@
   "Message": "Порука",
   "Mini QR Code Generator": "Мини генератор QR кодова",
   "MiniQR's QR rendering engine has been updated": "Мотор за приказивање QR кодова компаније MiniQR је ажуриран.",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Већина QR скенера неће прочитати овај формат јер симболи # остављају празнине између редова. Користите Уникод (полублокове/пуне блокове) за скењив извоз.",
   "Must be between 0 and 1": "Мора бити између 0 и 1",
   "Network Name (SSID)": "Име мреже (SSID)",
   "Network password (case-sensitive)": "Лозинка за мрежу (осетљива на велика и мала слова)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Молимо вас да прво отпремите фајл.",
   "Position": "Позиција",
   "Preset": "Претходно подешавање",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Преглед првог реда",
   "Processing...": "Обрада...",
   "QR Code Content": "Садржај QR кода",
   "QR code preset": "QR код пресет",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Једноставан URL/Текст",
   "Single export": "Појединачни извоз",
   "SMS": "СМС",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Чврсти блоковни знакови (██) — најбоља верност у савременим фонтовима.",
   "Sponsor": "Спонзор",
   "square": "квадрат",
   "Start Date": "Датум почетка",
@@ -291,15 +291,15 @@
   "Text": "Текст",
   "Text color": "Боја текста",
   "Text is required": "Текст је обавезан",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Испис текста искључује оквир и слику.",
   "Text position": "Позиција текста",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Превише QR кодова се истовремено копира у међуспремник. Молимо преузмите их као SVG или PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Савет: За најбоље резултате користите јасну слику са добрим осветљењем.",
   "Title": "Наслов",
   "Toggle dark mode": "Пребаци мрачни режим",
   "top": "најбоље",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Јуникод (цели блокови)",
+  "Unicode (half-blocks)": "Јуникод (полублокови)",
   "Upload a CSV file": "Отпремите CSV фајл",
   "Upload image": "Отпреми слику",
   "Upload QR Code Image": "Отпреми слику QR кода",
