+++ locales/sv.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Laddad från lokal lagring",
   "{count} piece(s) of data detected": "{count} del(ar) av data upptäckta",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR-koder har skapats.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-bitars tecken – fungerar i ren text, e-post och terminaler.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bitars, inverterad polaritet (# är bakgrundsfärg) — fungerar i ren text, e-post och terminaler.",
   "Add frame": "Lägg till ram",
   "Address": "Adress",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Stäng dialogrutan",
   "Close guide": "Stäng guide",
   "Close scanner": "Stäng scanner",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompakt: två rader sammanpressade till en rad med hjälp av tecknen ▀ ▄ █.",
   "Configure frame settings": "Konfigurera inställningar för ram",
   "Contact Card": "Kontaktkort",
-  "Copied": "Copied",
+  "Copied": "Kopierat",
   "Copy QR Code to clipboard": "Kopiera QR-kod till urklipp",
   "Copy to clipboard": "Kopiera till urklipp",
   "Corners Dot color": "Hörn prickfärg",
@@ -113,10 +113,10 @@
   "dots": "prickar",
   "Dots color": "Prickar färg",
   "Dots type": "Typ av prickar",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Ladda ner .md",
+  "Download .txt": "Ladda ner .txt",
+  "Download all as .md": "Ladda ner allt som .md",
+  "Download all as .txt": "Ladda ner allt som .txt",
   "Download QR Code as JPG": "Ladda ner QR-koden som JPG",
   "Download QR Code as PNG": "Ladda ner QR-koden som PNG",
   "Download QR Code as SVG": "Ladda ner QR-kod som SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Exempel på fil (vCard)",
   "Example with Custom Filenames": "Exempel med anpassade filnamn",
   "Export": "Exportera",
-  "Export as text": "Export as text",
+  "Export as text": "Exportera som text",
   "Export QR code": "Exportera QR-kod",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Exportera QR-koden som ASCII- eller Unicode-text",
   "extra-rounded": "extra-rundad",
   "Failed to copy to clipboard": "Det gick inte att kopiera till urklipp",
   "Failed to load changelog": "Misslyckades med att ladda changelog",
@@ -161,12 +161,12 @@
   "First Name": "Förnamn",
   "First row preview:": "Förhandsgranskning av första raden:",
   "Font family": "Teckensnittsfamilj",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Rametiketten ingår inte i textexporten – ASCII-utdata innehåller endast QR-modulerna.",
   "Frame preset": "Förinställd ram",
   "Frame settings": "Inställningar för ram",
   "Frame style": "Ram stil",
   "Frame text": "Ramtext",
-  "Generating": "Generating",
+  "Generating": "Generera",
   "GitHub repository for this project": "GitHub utvecklingskatalog för detta projekt",
   "Have a beautiful day!": "Ha en vacker dag!",
   "Have nice day!": "Ha en trevlig dag!",
@@ -204,7 +204,7 @@
   "Message": "Meddelande",
   "Mini QR Code Generator": "Mini QR-kodgenerator",
   "MiniQR's QR rendering engine has been updated": "MiniQRs QR-renderingmotor har uppdaterats",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "De flesta QR-läsare kan inte läsa detta format eftersom #-tecknen skapar mellanrum mellan raderna. Använd Unicode (halv-/helblock) för en export som går att skanna.",
   "Must be between 0 and 1": "Måste ligga mellan 0 och 1",
   "Network Name (SSID)": "Nätverksnamn (SSID)",
   "Network password (case-sensitive)": "Nätverkslösenord (skiftlägeskänsligt)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Ladda upp en fil först",
   "Position": "Befattning",
   "Preset": "Förinställd",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Förhandsgranskning av första raden",
   "Processing...": "Bearbetar...",
   "QR Code Content": "QR-kodens innehåll",
   "QR code preset": "Förinställd QR-kod",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Enkel URL/text",
   "Single export": "Enkel export",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Fyllda tecken (██) — bästa återgivningen i moderna typsnitt.",
   "Sponsor": "Sponsorn",
   "square": "kvadrat",
   "Start Date": "Startdatum",
@@ -291,15 +291,15 @@
   "Text": "Text",
   "Text color": "Text färg",
   "Text is required": "Text krävs",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Textutmatningen omfattar inte ramen och bilden",
   "Text position": "Textposition",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Det finns för många QR-koder för att kopieras till urklipp samtidigt. Ladda ner dem som SVG eller PNG istället.",
   "Tip: For best results, use a clear image with good lighting.": "Tips: För bästa resultat, använd en tydlig bild med bra belysning.",
   "Title": "Titel",
   "Toggle dark mode": "Växla mörkt läge",
   "top": "topp",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (hela block)",
+  "Unicode (half-blocks)": "Unicode (halvblock)",
   "Upload a CSV file": "Ladda upp en CSV-fil",
   "Upload image": "Ladda upp bild",
   "Upload QR Code Image": "Ladda upp QR-kod bild",
