+++ locales/th.json
@@ -51,11 +51,11 @@
   "Last saved locally": "อ่านข้อมูลจากพื้นที่จัดเก็บในเครื่อง",
   "{count} piece(s) of data detected": "{count} ชิ้นของข้อมูลที่ถูกพบ",
   "{index} / {count} QR codes have been created.": "สร้าง QR Code แล้ว {index} / {count} ",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "ตัวอักษร 7 บิต — ใช้งานได้ในข้อความธรรมดา อีเมล และเทอร์มินัล",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 บิต, พลิกขั้ว (# เป็นพื้นหลัง) — ใช้งานได้ในข้อความธรรมดา, อีเมล, และเทอร์มินัล",
   "Add frame": "เพิ่มกรอบรูป",
   "Address": "ที่อยู่",
-  "ASCII": "ASCII",
+  "ASCII": "เอสไควี (ASCII)",
   "Available Fields": "ฟิลด์ที่มีอยู่",
   "Background color": "สีพื้นหลัง",
   "Batch export": "ส่งออกเป็นชุด",
@@ -82,10 +82,10 @@
   "Close dialog": "ปิดกล่องโต้ตอบ",
   "Close guide": "ปิดคู่มือ",
   "Close scanner": "ปิดเครื่องสแกน",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "กะทัดรัด: สองแถวบรรจุในหนึ่งบรรทัดโดยใช้ตัวอักษร ▀ ▄ █",
   "Configure frame settings": "ตั้งค่ากรอบ",
   "Contact Card": "บัตรติดต่อ",
-  "Copied": "Copied",
+  "Copied": "คัดลอกแล้ว",
   "Copy QR Code to clipboard": "คัดลอก QR Code ไปยังคลิปบอร์ด",
   "Copy to clipboard": "คัดลอกไปยังคลิปบอร์ด",
   "Corners Dot color": "สีจุดมุม",
@@ -113,10 +113,10 @@
   "dots": "จุด",
   "Dots color": "สีจุด",
   "Dots type": "ประเภทจุด",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "ดาวน์โหลด .md",
+  "Download .txt": "ดาวน์โหลด .txt",
+  "Download all as .md": "ดาวน์โหลดทั้งหมดเป็น .md",
+  "Download all as .txt": "ดาวน์โหลดทั้งหมดเป็น .txt",
   "Download QR Code as JPG": "ดาวน์โหลด QR Code เป็น JPG",
   "Download QR Code as PNG": "ดาวน์โหลด QR Code เป็น PNG",
   "Download QR Code as SVG": "ดาวน์โหลด QR Code เป็น SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "ไฟล์ตัวอย่าง (vCard)",
   "Example with Custom Filenames": "ตัวอย่างพร้อมชื่อไฟล์ที่กำหนดเอง",
   "Export": "ส่งออก",
-  "Export as text": "Export as text",
+  "Export as text": "ส่งออกเป็นข้อความ",
   "Export QR code": "ส่งออก QR code",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "ส่งออก QR Code เป็นข้อความ ASCII หรือ Unicode",
   "extra-rounded": "กลมพิเศษ",
   "Failed to copy to clipboard": "ไม่สามารถคัดลอกไปยังคลิปบอร์ดได้",
   "Failed to load changelog": "ไม่สามารถโหลดประวัติการเปลี่ยนแปลงได้",
@@ -161,12 +161,12 @@
   "First Name": "ชื่อ",
   "First row preview:": "แสดงตัวอย่างแถวแรก:",
   "Font family": "ตระกูลฟอนต์",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "ป้ายชื่อเฟรมจะไม่ถูกรวมอยู่ในการส่งออกข้อความ — ผลลัพธ์ในรูปแบบ ASCII จะประกอบด้วยโมดูล QR เท่านั้น",
   "Frame preset": "เฟรมพรีเซ็ต",
   "Frame settings": "การตั้งค่ากรอบรูป",
   "Frame style": "รูปแบบกรอบรูป",
   "Frame text": "ข้อความในกรอบ",
-  "Generating": "Generating",
+  "Generating": "การสร้าง",
   "GitHub repository for this project": "GitHub repository สำหรับโปรเจ็กต์นี้",
   "Have a beautiful day!": "ขอให้มีวันที่สวยงาม!",
   "Have nice day!": "ขอให้มีวันที่ดี!",
@@ -204,7 +204,7 @@
   "Message": "ข้อความ",
   "Mini QR Code Generator": "มินิ QR โค้ด เจเนอเรเตอร์",
   "MiniQR's QR rendering engine has been updated": "ระบบแสดงผล QR ของ MiniQR ได้รับการอัปเดตแล้ว",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "เครื่องสแกน QR ส่วนใหญ่จะไม่สามารถอ่านรูปแบบนี้ได้ เนื่องจากสัญลักษณ์ # ทำให้มีช่องว่างระหว่างแถว ใช้ Unicode (บล็อกครึ่ง/เต็ม) สำหรับการส่งออกที่สามารถสแกนได้",
   "Must be between 0 and 1": "ต้องอยู่ระหว่าง 0 ถึง 1",
   "Network Name (SSID)": "ชื่อเครือข่าย (SSID)",
   "Network password (case-sensitive)": "รหัสผ่านเครือข่าย (พิมพ์เล็กพิมพ์ใหญ่มีผล)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "โปรดอัปโหลดไฟล์ก่อน",
   "Position": "ตำแหน่ง",
   "Preset": "ค่าตั้งต้น",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "ตัวอย่างแถวแรก",
   "Processing...": "กำลังดำเนินการ...",
   "QR Code Content": "เนื้อหา QR Code",
   "QR code preset": "ค่าตั้งต้น QR Code",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL/ข้อความง่าย ๆ",
   "Single export": "ส่งออกเดี่ยว",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "ตัวอักษรบล็อกทึบ (██) — ความเที่ยงตรงสูงสุดในฟอนต์สมัยใหม่",
   "Sponsor": "ผู้สนับสนุน",
   "square": "สี่เหลี่ยม",
   "Start Date": "วันที่เริ่มต้น",
@@ -291,15 +291,15 @@
   "Text": "ข้อความ",
   "Text color": "สีข้อความ",
   "Text is required": "ต้องระบุข้อความ",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "ข้อความที่แสดงผลไม่รวมกรอบและรูปภาพ",
   "Text position": "ตำแหน่งข้อความ",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "มี QR Code จำนวนมากเกินไปที่จะคัดลอกไปยังคลิปบอร์ดพร้อมกัน โปรดดาวน์โหลดเป็น SVG หรือ PNG แทน",
   "Tip: For best results, use a clear image with good lighting.": "คำแนะนำ: เพื่อผลลัพธ์ที่ดีที่สุด ควรใช้ภาพที่คมชัดและมีแสงสว่างเพียงพอ",
   "Title": "ตำแหน่ง",
   "Toggle dark mode": "สลับโหมดมืด",
   "top": "ด้านบน",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "ยูนิโค้ด (บล็อกเต็ม)",
+  "Unicode (half-blocks)": "ยูนิโค้ด (ครึ่งบล็อก)",
   "Upload a CSV file": "อัปโหลดไฟล์ CSV",
   "Upload image": "อัพโหลดรูปภาพ",
   "Upload QR Code Image": "อัปโหลดรูปภาพ QR Code",
