+++ locales/tr.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Yerel depodan yüklendi",
   "{count} piece(s) of data detected": "{count} tespit edilen veri parçası(ları)",
   "{index} / {count} QR codes have been created.": "{index} / {count} QR kodları oluşturulmuştur.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7 bitlik karakterler — düz metin, e-posta ve terminallerde çalışır.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 bit, ters polarite (# arka plan rengidir) — düz metin, e-posta ve terminallerde çalışır.",
   "Add frame": "Çerçeve ekleyin",
   "Address": "Adres",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "İletişim kutusunu kapat",
   "Close guide": "Kılavuzu kapat",
   "Close scanner": "Tarayıcıyı kapat",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Kompakt: ▀ ▄ █ karakterleri kullanılarak iki sıra tek bir satıra sığdırılmış.",
   "Configure frame settings": "Çerçeve ayarlarını yapılandırma",
   "Contact Card": "İletişim Kartı",
-  "Copied": "Copied",
+  "Copied": "Kopyalandı",
   "Copy QR Code to clipboard": "QR Kodunu panoya kopyalama",
   "Copy to clipboard": "Panoya kopyala",
   "Corners Dot color": "Köşeler Nokta rengi",
@@ -113,10 +113,10 @@
   "dots": "noktalar",
   "Dots color": "Nokta rengi",
   "Dots type": "Nokta tipi",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": ".md dosyasını indir",
+  "Download .txt": ".txt dosyasını indir",
+  "Download all as .md": "Hepsini .md olarak indir",
+  "Download all as .txt": "Hepsini .txt olarak indir",
   "Download QR Code as JPG": "QR Kodunu JPG Olarak İndirin",
   "Download QR Code as PNG": "QR Kodunu PNG Olarak İndirin",
   "Download QR Code as SVG": "QR Kodunu SVG Olarak İndirin",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Örnek dosya (vCard)",
   "Example with Custom Filenames": "Özel Dosya Adları ile Örnek",
   "Export": "Dışa aktar",
-  "Export as text": "Export as text",
+  "Export as text": "Metin olarak dışa aktar",
   "Export QR code": "QR kodunu dışa aktar",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "QR Kodunu ASCII veya Unicode metni olarak dışa aktar",
   "extra-rounded": "ekstra yuvarlak",
   "Failed to copy to clipboard": "Panoya kopyalanamadı",
   "Failed to load changelog": "Değişiklik günlüğü yüklenemedi",
@@ -161,12 +161,12 @@
   "First Name": "İlk İsim",
   "First row preview:": "İlk sıra önizlemesi:",
   "Font family": "Yazı tipi ailesi",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Çerçeve etiketi metin dışa aktarımından hariç tutulur — ASCII çıktısı yalnızca QR modüllerini içerir.",
   "Frame preset": "Çerçeve ön ayarı",
   "Frame settings": "Çerçeve ayarları",
   "Frame style": "Çerçeve stili",
   "Frame text": "Çerçeve metni",
-  "Generating": "Generating",
+  "Generating": "Oluşturma",
   "GitHub repository for this project": "Bu proje için GitHub deposu",
   "Have a beautiful day!": "Güzel bir gün geçirin!",
   "Have nice day!": "İyi günler!",
@@ -204,7 +204,7 @@
   "Message": "Mesaj",
   "Mini QR Code Generator": "Mini QR Kod Oluşturucu",
   "MiniQR's QR rendering engine has been updated": "MiniQR'ın QR görüntüleme motoru güncellendi",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Çoğu QR tarayıcı, # karakterlerinin satırlar arasında boşluk bırakması nedeniyle bu biçimi okumaz. Taranabilir bir dışa aktarım için Unicode (yarım/tam bloklar) kullanın.",
   "Must be between 0 and 1": "0 ile 1 arasında olmalıdır",
   "Network Name (SSID)": "Ağ Adı (SSID)",
   "Network password (case-sensitive)": "Ağ parolası (büyük/küçük harfe duyarlı)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Lütfen önce bir dosya yükleyin",
   "Position": "Pozisyon",
   "Preset": "Ön ayar",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "İlk satırın önizlemesi",
   "Processing...": "İşlem...",
   "QR Code Content": "QR Kod İçeriği",
   "QR code preset": "QR kodu ön ayarı",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Basit URL/Metin",
   "Single export": "Tek ihracat",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Düz blok karakterler (██) — modern yazı tiplerinde en yüksek kalitede görüntülenir.",
   "Sponsor": "Sponsor",
   "square": "kare",
   "Start Date": "Başlangıç Tarihi",
@@ -291,15 +291,15 @@
   "Text": "Metin",
   "Text color": "Metin rengi",
   "Text is required": "Metin gereklidir",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Metin çıktısı, çerçeve ve görüntüyü içermez",
   "Text position": "Metin konumu",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Panoya bir kerede kopyalanamayacak kadar çok QR kodu var. Lütfen bunun yerine SVG veya PNG olarak indirin.",
   "Tip: For best results, use a clear image with good lighting.": "İpucu: En iyi sonuçları elde etmek için iyi aydınlatılmış net bir görüntü kullanın.",
   "Title": "Başlık",
   "Toggle dark mode": "Karanlık modu aç/kapat",
   "top": "üst",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (tam bloklar)",
+  "Unicode (half-blocks)": "Unicode (yarım bloklar)",
   "Upload a CSV file": "CSV dosyası yükleme",
   "Upload image": "Resim yükle",
   "Upload QR Code Image": "QR Kod Görüntüsü Yükleme",
