+++ locales/uk.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Завантажено з локальної пам'яті",
   "{count} piece(s) of data detected": "Виявлено {count} одиниць даних",
   "{index} / {count} QR codes have been created.": "Створено QR-коди {index} / {count}.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7-бітні символи — підтримуються у звичайному тексті, електронній пошті та терміналах.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-бітний, з інвертованою полярністю (символ # — це фон) — працює у звичайному тексті, електронній пошті та терміналах.",
   "Add frame": "Додати фрейм",
   "Address": "Адреса",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Закрити діалогове вікно",
   "Close guide": "Закрити довідник",
   "Close scanner": "Закрити сканер",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Компактний варіант: два ряди, об’єднані в один за допомогою символів ▀ ▄ █.",
   "Configure frame settings": "Налаштуйте параметри кадру",
   "Contact Card": "Картка контакту",
-  "Copied": "Copied",
+  "Copied": "Скопійовано",
   "Copy QR Code to clipboard": "Скопіювати QR-код в буфер обміну",
   "Copy to clipboard": "Скопіювати до буферу обміну",
   "Corners Dot color": "Колір підкреслення кутів",
@@ -113,10 +113,10 @@
   "dots": "крапки",
   "Dots color": "Колір точок",
   "Dots type": "Тип крапок",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Завантажити .md",
+  "Download .txt": "Завантажити .txt",
+  "Download all as .md": "Завантажити все у форматі .md",
+  "Download all as .txt": "Завантажити все у форматі .txt",
   "Download QR Code as JPG": "Завантажити QR-код у форматі JPG",
   "Download QR Code as PNG": "Завантажити QR-код в форматі PNG",
   "Download QR Code as SVG": "Завантажити QR-код як SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Приклад файлу (vCard)",
   "Example with Custom Filenames": "Приклад із власними іменами файлів",
   "Export": "Експорт",
-  "Export as text": "Export as text",
+  "Export as text": "Експортувати як текст",
   "Export QR code": "Експорт QR-коду",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Експортувати QR-код у вигляді тексту у форматі ASCII або Unicode",
   "extra-rounded": "поза межами округлення",
   "Failed to copy to clipboard": "Не вдалося скопіювати в буфер обміну",
   "Failed to load changelog": "Не вдалося завантажити журнал змін",
@@ -161,12 +161,12 @@
   "First Name": "Ім'я та прізвище",
   "First row preview:": "Попередній перегляд першого рядка:",
   "Font family": "Сімейство шрифтів",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Етикетка кадру не включається в експорт тексту — вихідні дані у форматі ASCII містять лише QR-модулі.",
   "Frame preset": "Попереднє налаштування кадру",
   "Frame settings": "Налаштування рамки",
   "Frame style": "Тип рамки",
   "Frame text": "Текст кадру",
-  "Generating": "Generating",
+  "Generating": "Створення",
   "GitHub repository for this project": "GitHub репозиторій для цього проекту",
   "Have a beautiful day!": "Гарного вам дня!",
   "Have nice day!": "Гарного дня!",
@@ -204,7 +204,7 @@
   "Message": "Повідомлення",
   "Mini QR Code Generator": "Генератор міні-QR-кодів",
   "MiniQR's QR rendering engine has been updated": "Оновлено механізм візуалізації QR-кодів MiniQR",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Більшість QR-сканерів не зможуть зчитати цей формат, оскільки символи # залишають пробіли між рядками. Для експорту, придатного для сканування, використовуйте Unicode (напівблоки/повні блоки).",
   "Must be between 0 and 1": "Повинно бути в діапазоні від 0 до 1",
   "Network Name (SSID)": "Ім'я мережі (SSID)",
   "Network password (case-sensitive)": "Мережевий пароль (з урахуванням регістру)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Будь ласка, спочатку завантажте файл",
   "Position": "Позиція",
   "Preset": "Пресет",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Попередній перегляд першого рядка",
   "Processing...": "Йде обробка...",
   "QR Code Content": "Зміст QR-коду",
   "QR code preset": "QR-код за замовчуванням",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "Простий URL/текст",
   "Single export": "Єдиний експорт",
   "SMS": "SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Суцільні символи (██) — найкраща якість відтворення в сучасних шрифтах.",
   "Sponsor": "Спонсорство",
   "square": "квадрат",
   "Start Date": "Дата початку",
@@ -291,15 +291,15 @@
   "Text": "Текст",
   "Text color": "Колір тексту",
   "Text is required": "Текст обов'язковий",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "У текстовому виведенні не відображаються рамка та зображення",
   "Text position": "Положення тексту",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Необхідно скопіювати забагато QR-кодів в буфер обміну одноразово. Будь ласка, завантажте їх через SVG або PNG.",
   "Tip: For best results, use a clear image with good lighting.": "Підказка: щоб отримати найкращі результати, використовуйте чисте зображення з хорошим освітленням.",
   "Title": "Назва",
   "Toggle dark mode": "Перемкнути темний режим",
   "top": "згори",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (повні блоки)",
+  "Unicode (half-blocks)": "Unicode (напівблоки)",
   "Upload a CSV file": "Завантажте файл CSV",
   "Upload image": "Вивантажити зображення",
   "Upload QR Code Image": "Завантажити зображення QR-коду",
