+++ locales/vi.json
@@ -51,8 +51,8 @@
   "Last saved locally": "Được tải từ bộ nhớ cục bộ",
   "{count} piece(s) of data detected": "{count} Dữ liệu được phát hiện (các phần dữ liệu)",
   "{index} / {count} QR codes have been created.": "{index} / {count} Mã QR đã được tạo.",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "Ký tự 7-bit — hoạt động trong văn bản thuần túy, email và các thiết bị đầu cuối.",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, cực tính đảo ngược (# là màu nền) — hoạt động trong văn bản thuần túy, email và các thiết bị đầu cuối.",
   "Add frame": "Thêm khung",
   "Address": "Địa chỉ",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "Đóng hộp thoại",
   "Close guide": "Hướng dẫn chi tiết",
   "Close scanner": "Tắt máy quét",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Dạng gọn: hai hàng được xếp gọn thành một dòng bằng các ký tự ▀ ▄ █.",
   "Configure frame settings": "Cấu hình cài đặt khung",
   "Contact Card": "Thẻ liên hệ",
-  "Copied": "Copied",
+  "Copied": "Đã sao chép",
   "Copy QR Code to clipboard": "Sao chép mã QR vào khay nhớ tạm",
   "Copy to clipboard": "Sao chép vào khay nhớ tạm",
   "Corners Dot color": "Màu chấm góc",
@@ -113,10 +113,10 @@
   "dots": "chấm",
   "Dots color": "Màu chấm",
   "Dots type": "Loại chấm",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "Tải xuống tệp .md",
+  "Download .txt": "Tải xuống tệp .txt",
+  "Download all as .md": "Tải xuống tất cả dưới dạng .md",
+  "Download all as .txt": "Tải xuống tất cả dưới dạng tệp .txt",
   "Download QR Code as JPG": "Tải xuống mã QR dưới dạng tệp JPG",
   "Download QR Code as PNG": "Tải xuống mã QR dưới dạng PNG",
   "Download QR Code as SVG": "Tải xuống mã QR dưới dạng SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "Tệp mẫu (vCard)",
   "Example with Custom Filenames": "Ví dụ với tên tệp tùy chỉnh",
   "Export": "Xuất khẩu",
-  "Export as text": "Export as text",
+  "Export as text": "Xuất dưới dạng văn bản",
   "Export QR code": "Xuất mã QR",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "Xuất mã QR dưới dạng văn bản ASCII hoặc Unicode",
   "extra-rounded": "tròn hơn",
   "Failed to copy to clipboard": "Không thể sao chép vào khay nhớ tạm.",
   "Failed to load changelog": "Không thể tải nhật ký thay đổi",
@@ -161,12 +161,12 @@
   "First Name": "Họ",
   "First row preview:": "Xem trước hàng đầu tiên:",
   "Font family": "Gia đình phông chữ",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Nhãn khung không được bao gồm trong tệp văn bản xuất ra — tệp ASCII xuất ra chỉ chứa các mô-đun QR.",
   "Frame preset": "Cài đặt khung mặc định",
   "Frame settings": "Cài đặt khung",
   "Frame style": "Kiểu khung",
   "Frame text": "Khung văn bản",
-  "Generating": "Generating",
+  "Generating": "Tạo",
   "GitHub repository for this project": "Kho lưu trữ GitHub cho dự án này",
   "Have a beautiful day!": "Chúc bạn có một ngày thật đẹp!",
   "Have nice day!": "Chúc bạn một ngày tốt lành!",
@@ -204,7 +204,7 @@
   "Message": "Thông điệp",
   "Mini QR Code Generator": "Trình tạo mã QR mini",
   "MiniQR's QR rendering engine has been updated": "Công cụ hiển thị mã QR của MiniQR đã được cập nhật",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Hầu hết các trình quét mã QR sẽ không đọc được định dạng này vì các ký tự # tạo ra khoảng trống giữa các hàng. Hãy sử dụng Unicode (khối nửa/đầy đủ) để xuất ra định dạng có thể quét được.",
   "Must be between 0 and 1": "Phải nằm trong khoảng từ 0 đến 1",
   "Network Name (SSID)": "Tên mạng (SSID)",
   "Network password (case-sensitive)": "Mật khẩu mạng (nhạy cảm với chữ hoa/thường)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "Vui lòng tải lên tệp tin trước tiên.",
   "Position": "Vị trí",
   "Preset": "Cài đặt sẵn",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "Xem trước hàng đầu tiên",
   "Processing...": "Đang xử lý...",
   "QR Code Content": "Nội dung mã QR",
   "QR code preset": "Mã QR đã được thiết lập sẵn",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "URL đơn giản/Văn bản",
   "Single export": "Xuất khẩu đơn lẻ",
   "SMS": "Tin nhắn SMS",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "Ký tự khối đặc (██) — độ trung thực cao nhất trong các phông chữ hiện đại.",
   "Sponsor": "Nhà tài trợ",
   "square": "hình vuông",
   "Start Date": "Ngày bắt đầu",
@@ -291,15 +291,15 @@
   "Text": "Văn bản",
   "Text color": "Màu văn bản",
   "Text is required": "Văn bản là bắt buộc.",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "Kết quả hiển thị không bao gồm khung và hình ảnh",
   "Text position": "Vị trí văn bản",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "Có quá nhiều mã QR để sao chép vào khay nhớ tạm cùng một lúc. Vui lòng tải chúng xuống dưới dạng SVG hoặc PNG thay thế.",
   "Tip: For best results, use a clear image with good lighting.": "Mẹo: Để đạt kết quả tốt nhất, hãy sử dụng hình ảnh rõ nét với ánh sáng tốt.",
   "Title": "Tiêu đề",
   "Toggle dark mode": "Bật/tắt chế độ tối",
   "top": "đỉnh",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode (khối đầy đủ)",
+  "Unicode (half-blocks)": "Unicode (khối nửa)",
   "Upload a CSV file": "Tải lên tệp CSV",
   "Upload image": "Tải lên hình ảnh",
   "Upload QR Code Image": "Tải lên hình ảnh mã QR",
