+++ locales/zh.json
@@ -51,8 +51,8 @@
   "Last saved locally": "從本機儲存載入",
   "{count} piece(s) of data detected": "{count} 偵測到的資料",
   "{index} / {count} QR codes have been created.": "{index} / {count} 已建立 QR 代碼。",
-  "7-bit characters — works in plain text, email, and terminals.": "7-bit characters — works in plain text, email, and terminals.",
-  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.",
+  "7-bit characters — works in plain text, email, and terminals.": "7 位元字元 — 適用於純文字、電子郵件及終端機。",
+  "7-bit, inverted polarity (# is background) — works in plain text, email, and terminals.": "7 位元，極性反轉（# 代表背景色）— 適用於純文字、電子郵件及終端機。",
   "Add frame": "新增外框",
   "Address": "地址",
   "ASCII": "ASCII",
@@ -82,10 +82,10 @@
   "Close dialog": "關閉對話框",
   "Close guide": "關閉指南",
   "Close scanner": "關閉掃描器",
-  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "Compact: two rows packed into one line using ▀ ▄ █ characters.",
+  "Compact: two rows packed into one line using ▀ ▄ █ characters.": "緊湊型：使用 ▀ ▄ █ 字元將兩行壓縮為一行。",
   "Configure frame settings": "設定框架設定",
   "Contact Card": "聯絡卡",
-  "Copied": "Copied",
+  "Copied": "已複製",
   "Copy QR Code to clipboard": "複製 QR 代碼到剪貼簿",
   "Copy to clipboard": "複製到剪貼簿",
   "Corners Dot color": "角點顏色",
@@ -113,10 +113,10 @@
   "dots": "小點",
   "Dots color": "圓點顏色",
   "Dots type": "圓點類型",
-  "Download .md": "Download .md",
-  "Download .txt": "Download .txt",
-  "Download all as .md": "Download all as .md",
-  "Download all as .txt": "Download all as .txt",
+  "Download .md": "下載 .md",
+  "Download .txt": "下載 .txt 檔案",
+  "Download all as .md": "將所有內容下載為 .md 格式",
+  "Download all as .txt": "將所有內容下載為 .txt 檔案",
   "Download QR Code as JPG": "下載 QR 代碼為 JPG",
   "Download QR Code as PNG": "下載 QR Code 為 PNG",
   "Download QR Code as SVG": "下載 QR Code 為 SVG",
@@ -149,9 +149,9 @@
   "Example file (vCard)": "範例檔案 (vCard)",
   "Example with Custom Filenames": "自訂檔案名稱範例",
   "Export": "出口",
-  "Export as text": "Export as text",
+  "Export as text": "匯出為文字",
   "Export QR code": "匯出 QR 碼",
-  "Export QR Code as ASCII or Unicode text": "Export QR Code as ASCII or Unicode text",
+  "Export QR Code as ASCII or Unicode text": "將 QR 碼匯出為 ASCII 或 Unicode 文字",
   "extra-rounded": "特圓",
   "Failed to copy to clipboard": "複製到剪貼簿失敗",
   "Failed to load changelog": "載入更新記錄失敗",
@@ -161,12 +161,12 @@
   "First Name": "姓名",
   "First row preview:": "第一排預覽：",
   "Font family": "字型家族",
-  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "Frame label is excluded from text export — the ASCII output contains the QR modules only.",
+  "Frame label is excluded from text export — the ASCII output contains the QR modules only.": "幀標籤不包含在文字匯出中 — ASCII 輸出僅包含 QR 模組。",
   "Frame preset": "框架預設",
   "Frame settings": "框架設定",
   "Frame style": "外框樣式",
   "Frame text": "框架文字",
-  "Generating": "Generating",
+  "Generating": "生成",
   "GitHub repository for this project": "本專案的 GitHub 套件庫",
   "Have a beautiful day!": "祝您有美好的一天",
   "Have nice day!": "祝您愉快",
@@ -204,7 +204,7 @@
   "Message": "訊息",
   "Mini QR Code Generator": "迷你 QR 代碼產生器",
   "MiniQR's QR rendering engine has been updated": "MiniQR 的 QR 碼渲染引擎已更新",
-  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.",
+  "Most QR scanners will not read this format because the # glyphs leave gaps between rows. Use Unicode (half/full blocks) for a scannable export.": "大多數 QR 碼掃描器無法讀取此格式，因為 # 字元會在行與行之間留下空格。若要產生可掃描的輸出檔，請使用 Unicode（半形/全形字塊）。",
   "Must be between 0 and 1": "必須介於 0 和 1 之間",
   "Network Name (SSID)": "網路名稱 (SSID)",
   "Network password (case-sensitive)": "網路密碼 (區分大小寫)",
@@ -243,7 +243,7 @@
   "Please upload a file first": "請先上傳檔案",
   "Position": "職位",
   "Preset": "預設",
-  "Preview of first row": "Preview of first row",
+  "Preview of first row": "第一行的預覽",
   "Processing...": "處理...",
   "QR Code Content": "QR 代碼內容",
   "QR code preset": "預設 QR 代碼",
@@ -274,7 +274,7 @@
   "Simple URL/Text": "簡單的 URL/文字",
   "Single export": "單一出口",
   "SMS": "簡訊",
-  "Solid block characters (██) — best fidelity in modern fonts.": "Solid block characters (██) — best fidelity in modern fonts.",
+  "Solid block characters (██) — best fidelity in modern fonts.": "實心方塊字元 (██) —— 在現代字型中呈現效果最佳。",
   "Sponsor": "贊助",
   "square": "方形",
   "Start Date": "開始日期",
@@ -291,15 +291,15 @@
   "Text": "正文",
   "Text color": "文字顏色",
   "Text is required": "必須使用文字",
-  "Text output excludes the frame and image": "Text output excludes the frame and image",
+  "Text output excludes the frame and image": "文字輸出不包含框架和圖片",
   "Text position": "文字位置",
   "There are too many QR codes to be copied to the clipboard at once. Please download them as SVG or PNG instead.": "QR 代碼太多，無法一次複製到剪貼簿。請下載為 SVG 或 PNG。",
   "Tip: For best results, use a clear image with good lighting.": "提示：為達到最佳效果，請使用光線良好的清晰影像。",
   "Title": "標題",
   "Toggle dark mode": "切換黑暗模式",
   "top": "最著名的",
-  "Unicode (full-blocks)": "Unicode (full-blocks)",
-  "Unicode (half-blocks)": "Unicode (half-blocks)",
+  "Unicode (full-blocks)": "Unicode（完整區塊）",
+  "Unicode (half-blocks)": "Unicode（半形區塊）",
   "Upload a CSV file": "上傳 CSV 檔案",
   "Upload image": "上傳圖片",
   "Upload QR Code Image": "上傳 QR 碼影像",
